Calculates lda dimension estimates

step_lda creates a specification of a recipe step that will return the lda dimension estimates of a text variable.

step_lda(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  lda_models = NULL,
  num_topics = 10L,
  prefix = "lda",
  skip = FALSE,
  id = rand_id("lda")
)

Source

https://arxiv.org/abs/1301.3781

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose which variables are affected by the step. See recipes::selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

columns

A character string of variable names that will be populated (eventually) by the terms argument. This is NULL until the step is trained by recipes::prep.recipe().

lda_models

A WarpLDA model object from the text2vec package. If left to NULL, the default, will it train its model based on the training data. Look at the examples for how to fit a WarpLDA model.

num_topics

integer desired number of latent topics.

prefix

A prefix for generated column names, default to "lda".

skip

A logical. Should the step be skipped when the recipe is baked by recipes::bake.recipe()? While all operations are baked when recipes::prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = FALSE.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

See also

Examples

if (requireNamespace("text2vec", quietly = TRUE)) {
# \donttest{
library(recipes)
library(modeldata)
data(tate_text)

tate_rec <- recipe(~ ., data = tate_text) %>%
  step_tokenize(medium) %>%
  step_lda(medium)

tate_obj <- tate_rec %>%
  prep()

bake(tate_obj, new_data = NULL) %>%
  slice(1:2)
tidy(tate_rec, number = 1)
tidy(tate_obj, number = 1)

# Changing the number of topics.
recipe(~ ., data = tate_text) %>%
  step_tokenize(medium, artist) %>%
  step_lda(medium, artist, num_topics = 20) %>%
  prep() %>%
  bake(new_data = NULL) %>%
  slice(1:2)

# Supplying A pre-trained LDA model trained using text2vec
library(text2vec)
tokens <- word_tokenizer(tolower(tate_text$medium))
it <- itoken(tokens, ids = seq_along(tate_text$medium))
v <- create_vocabulary(it)
dtm <- create_dtm(it, vocab_vectorizer(v))
lda_model <- LDA$new(n_topics = 15)

recipe(~ ., data = tate_text) %>%
  step_tokenize(medium, artist) %>%
  step_lda(medium, artist, lda_models = lda_model) %>%
  prep() %>%
  bake(new_data = NULL) %>%
  slice(1:2)
# }
}
#> # A tibble: 2 × 33
#>      id title       year lda_medium_w1 lda_medium_w2 lda_medium_w3 lda_medium_w4
#>   <dbl> <fct>      <dbl>         <dbl>         <dbl>         <dbl>         <dbl>
#> 1 21926 Proposals…  1990             0        0.0143             0             0
#> 2 20472 Michael     1990             0        0                  0             0
#> # … with 26 more variables: lda_medium_w5 <dbl>, lda_medium_w6 <dbl>,
#> #   lda_medium_w7 <dbl>, lda_medium_w8 <dbl>, lda_medium_w9 <dbl>,
#> #   lda_medium_w10 <dbl>, lda_medium_w11 <dbl>, lda_medium_w12 <dbl>,
#> #   lda_medium_w13 <dbl>, lda_medium_w14 <dbl>, lda_medium_w15 <dbl>,
#> #   lda_artist_w1 <dbl>, lda_artist_w2 <dbl>, lda_artist_w3 <dbl>,
#> #   lda_artist_w4 <dbl>, lda_artist_w5 <dbl>, lda_artist_w6 <dbl>,
#> #   lda_artist_w7 <dbl>, lda_artist_w8 <dbl>, lda_artist_w9 <dbl>, …