Generate the basic set of text features

step_textfeature creates a specification of a recipe step that will extract a number of numeric features of a text column.

step_textfeature(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  extract_functions = textfeatures::count_functions,
  prefix = "textfeature",
  skip = FALSE,
  id = rand_id("textfeature")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose which variables are affected by the step. See recipes::selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

columns

A character string of variable names that will be populated (eventually) by the terms argument. This is NULL until the step is trained by recipes::prep.recipe().

extract_functions

A named list of feature extracting functions. default to count_functions from the textfeatures package. See details for more information.

prefix

A prefix for generated column names, default to "textfeature".

skip

A logical. Should the step be skipped when the recipe is baked by recipes::bake.recipe()? While all operations are baked when recipes::prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = FALSE.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

This step will take a character column and returns a number of numeric columns equal to the number of functions in the list passed to the extract_functions argument. The default is a list of functions from the textfeatures package.

All the functions passed to extract_functions must take a character vector as input and return a numeric vector of the same length, otherwise an error will be thrown.

See also

Examples

if (requireNamespace("textfeatures", quietly = TRUE)) {
  library(recipes)
  library(modeldata)
  data(tate_text)

  tate_rec <- recipe(~., data = tate_text) %>%
    step_textfeature(medium)

  tate_obj <- tate_rec %>%
    prep()

  bake(tate_obj, new_data = NULL) %>%
    slice(1:2)

  bake(tate_obj, new_data = NULL) %>%
    pull(textfeature_medium_n_words)

  tidy(tate_rec, number = 1)
  tidy(tate_obj, number = 1)

  # Using custom extraction functions
  nchar_round_10 <- function(x) round(nchar(x) / 10) * 10

  recipe(~., data = tate_text) %>%
    step_textfeature(medium,
      extract_functions = list(nchar10 = nchar_round_10)
    ) %>%
    prep() %>%
    bake(new_data = NULL)
}
#> # A tibble: 4,284 × 5
#>        id artist             title                      year textfeature_medium…
#>     <dbl> <fct>              <fct>                     <dbl>               <dbl>
#>  1  21926 Absalon            Proposals for a Habitat    1990                  60
#>  2  20472 Auerbach, Frank    Michael                    1990                  20
#>  3  20474 Auerbach, Frank    Geoffrey                   1990                  20
#>  4  20473 Auerbach, Frank    Jake                       1990                  20
#>  5  20513 Auerbach, Frank    To the Studios             1990                  20
#>  6  21389 Ayres, OBE Gillian Phaëthon                   1990                  20
#>  7 121187 Barlow, Phyllida   Untitled                   1990                  20
#>  8  19455 Baselitz, Georg    Green VIII                 1990                  20
#>  9  20938 Beattie, Basil     Present Bound              1990                  30
#> 10 105941 Beuys, Joseph      Joseph Beuys: A Private …  1990                  10
#> # … with 4,274 more rows