Generate the basic set of text features

step_tokenmerge creates a specification of a recipe step that will take multiple tokenlists and combine them into one tokenlist.

step_tokenmerge(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  prefix = "tokenmerge",
  skip = FALSE,
  id = rand_id("tokenmerge")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose which variables are affected by the step. See recipes::selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

columns

A character string of variable names that will be populated (eventually) by the terms argument. This is NULL until the step is trained by recipes::prep.recipe().

prefix

A prefix for generated column names, default to "tokenmerge".

skip

A logical. Should the step be skipped when the recipe is baked by recipes::bake.recipe()? While all operations are baked when recipes::prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = FALSE.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

See also

Examples

library(recipes)
library(modeldata)
data(tate_text)

tate_rec <- recipe(~., data = tate_text) %>%
  step_tokenize(medium, artist) %>%
  step_tokenmerge(medium, artist)

tate_obj <- tate_rec %>%
  prep()

bake(tate_obj, new_data = NULL)
#> # A tibble: 4,284 × 4
#>        id title                                                  year tokenmerge
#>     <dbl> <fct>                                                 <dbl>  <tknlist>
#>  1  21926 Proposals for a Habitat                                1990 [9 tokens]
#>  2  20472 Michael                                                1990 [5 tokens]
#>  3  20474 Geoffrey                                               1990 [5 tokens]
#>  4  20473 Jake                                                   1990 [5 tokens]
#>  5  20513 To the Studios                                         1990 [6 tokens]
#>  6  21389 Phaëthon                                               1990 [7 tokens]
#>  7 121187 Untitled                                               1990 [6 tokens]
#>  8  19455 Green VIII                                             1990 [5 tokens]
#>  9  20938 Present Bound                                          1990 [8 tokens]
#> 10 105941 Joseph Beuys: A Private Collection. A11 Artforum, Mu…  1990 [5 tokens]
#> # … with 4,274 more rows

tidy(tate_rec, number = 1)
#> # A tibble: 2 × 3
#>   terms  value id            
#>   <chr>  <chr> <chr>         
#> 1 medium NA    tokenize_vvFzF
#> 2 artist NA    tokenize_vvFzF
tidy(tate_obj, number = 1)
#> # A tibble: 2 × 3
#>   terms  value id            
#>   <chr>  <chr> <chr>         
#> 1 medium words tokenize_vvFzF
#> 2 artist words tokenize_vvFzF