Calculates word2vec dimension estimates

`step_word2vec` creates a *specification* of a recipe step that will return the word2vec dimension estimates of a text variable.

step_word2vec(recipe, ..., role = "predictor", trained = FALSE,
  columns = NULL, lda_models = NULL, num_topics = 10,
  prefix = "word2vec", skip = FALSE, id = rand_id("word2vec"))

# S3 method for step_word2vec
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose variables. For `step_word2vec`, this indicates the variables to be encoded into a list column. See [recipes::selections()] for more details. For the `tidy` method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the recipe has been baked.

columns
A list of tibble results that define the encoding. This is `NULL` until the step is trained by [recipes::prep.recipe()].

lda_models
A WarpLDA model object from the text2vec package. If left to NULL, the default, will it train its model based on the training data. Look at the examples for how to fit a WarpLDA model.

num_topics
integer desired number of latent topics.

prefix
A prefix for generated column names, default to "word2vec".

skip
A logical. Should the step be skipped when the recipe is baked by [recipes::bake.recipe()]? While all operations are baked when [recipes::prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it

x
A `step_word2vec` object.

Source

https://arxiv.org/abs/1301.3781

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any).

Examples

library(recipes)

data(okc_text)

okc_rec <- recipe(~ ., data = okc_text) %>%
  step_word2vec(essay0)

okc_obj <- okc_rec %>%
  prep(training = okc_text, retain = TRUE)

juice(okc_obj) %>%
  slice(1:2)

#> # A tibble: 2 x 19
#>   essay1 essay2 essay3 essay4 essay5 essay6 essay7 essay8 essay9
#>   <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct> 
#> 1 "i ju… "writ… "that… "musi… "frie… "roma… "usua… "i ha… "you'…
#> 2 eatin… "pick… "i lo… "non-… "dese… "ever… makin… this … "you'…
#> # … with 10 more variables: word2vec_essay0_w1 <dbl>, word2vec_essay0_w2 <dbl>,
#> #   word2vec_essay0_w3 <dbl>, word2vec_essay0_w4 <dbl>,
#> #   word2vec_essay0_w5 <dbl>, word2vec_essay0_w6 <dbl>,
#> #   word2vec_essay0_w7 <dbl>, word2vec_essay0_w8 <dbl>,
#> #   word2vec_essay0_w9 <dbl>, word2vec_essay0_w10 <dbl>

tidy(okc_rec, number = 1)

#> # A tibble: 1 x 3
#>   terms  num_topics id            
#>   <chr>  <lgl>      <chr>         
#> 1 essay0 NA         word2vec_8LSSy

tidy(okc_obj, number = 1)

#> # A tibble: 1 x 3
#>   terms  num_topics id            
#>   <chr>       <dbl> <chr>         
#> 1 essay0         10 word2vec_8LSSy

# Changing the number of topics.
recipe(~ ., data = okc_text) %>%
  step_word2vec(essay0, essay1, num_topics = 20) %>%
  prep() %>%
  juice() %>%
  slice(1:2)

#> # A tibble: 2 x 48
#>   essay2 essay3 essay4 essay5 essay6 essay7 essay8 essay9 word2vec_essay0…
#>   <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>             <dbl>
#> 1 "writ… "that… "musi… "frie… "roma… "usua… "i ha… "you'…           0.0854
#> 2 "pick… "i lo… "non-… "dese… "ever… makin… this … "you'…           0     
#> # … with 39 more variables: word2vec_essay0_w2 <dbl>, word2vec_essay0_w3 <dbl>,
#> #   word2vec_essay0_w4 <dbl>, word2vec_essay0_w5 <dbl>,
#> #   word2vec_essay0_w6 <dbl>, word2vec_essay0_w7 <dbl>,
#> #   word2vec_essay0_w8 <dbl>, word2vec_essay0_w9 <dbl>,
#> #   word2vec_essay0_w10 <dbl>, word2vec_essay0_w11 <dbl>,
#> #   word2vec_essay0_w12 <dbl>, word2vec_essay0_w13 <dbl>,
#> #   word2vec_essay0_w14 <dbl>, word2vec_essay0_w15 <dbl>,
#> #   word2vec_essay0_w16 <dbl>, word2vec_essay0_w17 <dbl>,
#> #   word2vec_essay0_w18 <dbl>, word2vec_essay0_w19 <dbl>,
#> #   word2vec_essay0_w20 <dbl>, word2vec_essay1_w1 <dbl>,
#> #   word2vec_essay1_w2 <dbl>, word2vec_essay1_w3 <dbl>,
#> #   word2vec_essay1_w4 <dbl>, word2vec_essay1_w5 <dbl>,
#> #   word2vec_essay1_w6 <dbl>, word2vec_essay1_w7 <dbl>,
#> #   word2vec_essay1_w8 <dbl>, word2vec_essay1_w9 <dbl>,
#> #   word2vec_essay1_w10 <dbl>, word2vec_essay1_w11 <dbl>,
#> #   word2vec_essay1_w12 <dbl>, word2vec_essay1_w13 <dbl>,
#> #   word2vec_essay1_w14 <dbl>, word2vec_essay1_w15 <dbl>,
#> #   word2vec_essay1_w16 <dbl>, word2vec_essay1_w17 <dbl>,
#> #   word2vec_essay1_w18 <dbl>, word2vec_essay1_w19 <dbl>,
#> #   word2vec_essay1_w20 <dbl>

# Supplying A pre-trained LDA model trained using text2vec
library(text2vec)
tokens <- word_tokenizer(tolower(okc_text$essay5))
it <- itoken(tokens, ids = seq_along(okc_text$essay5))
v <- create_vocabulary(it)
dtm <- create_dtm(it, vocab_vectorizer(v))
lda_model <- LDA$new(n_topics = 15)

recipe(~ ., data = okc_text) %>%
  step_word2vec(essay0, essay1, lda_models = lda_model) %>%
  prep() %>%
  juice() %>%
  slice(1:2)

#> # A tibble: 2 x 38
#>   essay2 essay3 essay4 essay5 essay6 essay7 essay8 essay9 word2vec_essay0…
#>   <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>             <dbl>
#> 1 "writ… "that… "musi… "frie… "roma… "usua… "i ha… "you'…           0.0244
#> 2 "pick… "i lo… "non-… "dese… "ever… makin… this … "you'…           0     
#> # … with 29 more variables: word2vec_essay0_w2 <dbl>, word2vec_essay0_w3 <dbl>,
#> #   word2vec_essay0_w4 <dbl>, word2vec_essay0_w5 <dbl>,
#> #   word2vec_essay0_w6 <dbl>, word2vec_essay0_w7 <dbl>,
#> #   word2vec_essay0_w8 <dbl>, word2vec_essay0_w9 <dbl>,
#> #   word2vec_essay0_w10 <dbl>, word2vec_essay0_w11 <dbl>,
#> #   word2vec_essay0_w12 <dbl>, word2vec_essay0_w13 <dbl>,
#> #   word2vec_essay0_w14 <dbl>, word2vec_essay0_w15 <dbl>,
#> #   word2vec_essay1_w1 <dbl>, word2vec_essay1_w2 <dbl>,
#> #   word2vec_essay1_w3 <dbl>, word2vec_essay1_w4 <dbl>,
#> #   word2vec_essay1_w5 <dbl>, word2vec_essay1_w6 <dbl>,
#> #   word2vec_essay1_w7 <dbl>, word2vec_essay1_w8 <dbl>,
#> #   word2vec_essay1_w9 <dbl>, word2vec_essay1_w10 <dbl>,
#> #   word2vec_essay1_w11 <dbl>, word2vec_essay1_w12 <dbl>,
#> #   word2vec_essay1_w13 <dbl>, word2vec_essay1_w14 <dbl>,
#> #   word2vec_essay1_w15 <dbl>