Tidy the Result of a Recipe

tidy will return a data frame that contains information regarding a recipe or operation within the recipe (when a tidy method for the operation exists). See recipes::tidy.recipe for more information.

Usage

# S3 method for step_clean_levels
tidy(x, ...)

# S3 method for step_clean_names
tidy(x, ...)

# S3 method for step_word_embeddings
tidy(x, ...)

# S3 method for step_texthash
tidy(x, ...)

# S3 method for step_dummy_hash
tidy(x, ...)

# S3 method for step_lda
tidy(x, ...)

# S3 method for step_lemma
tidy(x, ...)

# S3 method for step_ngram
tidy(x, ...)

# S3 method for step_pos_filter
tidy(x, ...)

# S3 method for step_sequence_onehot
tidy(x, ...)

# S3 method for step_stem
tidy(x, ...)

# S3 method for step_stopwords
tidy(x, ...)

# S3 method for step_text_normalization
tidy(x, ...)

# S3 method for step_textfeature
tidy(x, ...)

# S3 method for step_tf
tidy(x, ...)

# S3 method for step_tfidf
tidy(x, ...)

# S3 method for step_tokenfilter
tidy(x, ...)

# S3 method for step_tokenize
tidy(x, ...)

# S3 method for step_tokenize_bpe
tidy(x, ...)

# S3 method for step_tokenize_sentencepiece
tidy(x, ...)

# S3 method for step_tokenize_wordpiece
tidy(x, ...)

# S3 method for step_tokenmerge
tidy(x, ...)

# S3 method for step_untokenize
tidy(x, ...)

Arguments

x

A step_untokenize object.

...

Not currently used.