OkCupid Text Data

These are a sample of columns and users of OkCupid dating website. The data are from Kim and Escobedo-Land (2015). Permission to use this data set was explicitly granted by OkCupid. The data set contains 10 text fields filled out by users.

Source

Kim, A. Y., and A. Escobedo-Land. 2015. "OkCupid Data for Introductory Statistics and Data Science Courses." Journal of Statistics Education: An International Journal on the Teaching and Learning of Statistics.

Value

okc_text

a tibble

Examples

data(okc_text)
str(okc_text)

#> tibble [750 × 10] (S3: tbl_df/tbl/data.frame)
#>  $ essay0: chr [1:750] "\"twice upon a time there was a boy who died twice and lived happily\never after but that's another chapter liv"| __truncated__ "i'm chill and steady.<br />\ni'm a teacher &amp; musician.<br />\ni like playing outside, dislike school nights"| __truncated__ "ok, i really don't know what (if anything) i'm looking for. impress\nme. i am a singer, songwriter, poet, amate"| __truncated__ "i am a dreamer who is constantly trying to come to terms with\nreality. i have lived in quite a few places in m"| __truncated__ ...
#>  $ essay1: chr [1:750] "i just graduated from college and have entered the \"what the fuck\nnow?!\" stage of my life.<br />\n<br />\ni'"| __truncated__ "eating, sleeping, and pooping." "raising my son, sometimes doing music, switching careers, having\nfun!" "i just moved back to san francisco from hong kong. i'm working on\nimmersing myself in the city. i am enjoying "| __truncated__ ...
#>  $ essay2: chr [1:750] "writing. being a friend. drawing. sarcasm. remembering lyrics.\ndancing. sleeping." "picking up/imitating people's mannerisms/voice.<br />\nbeing rational and respectful.<br />\ncooking." "procrastinating. crap..i shouldn't say that, should i? i'll change\nit later. ha ha! i'm good at writing music "| __truncated__ "i've been told i'm good at listening, writing, analyzing dilemmas,\nstory-telling, and photography." ...
#>  $ essay3: chr [1:750] "that i am playing really loud music or that i have a tad bit more\nass than the average man" "i look like i'm 15, or, \"so, what are you?\"" "my smile, my eyes, my dimples, my little hands. and of course, my\noverall awesomeness!" "my fashion accessories and my desire to have a good time." ...
#>  $ essay4: chr [1:750] "music:<br />\nerykah badu. . .nina simone. . .camp lo. . .nirvana. . .the coup. .\n.nas. . .biggie. . .lil kim."| __truncated__ "non-fiction.<br />\nchick flicks and trippy shit.<br />\nsurvivor.<br />\ni try hard to understand what's good "| __truncated__ "sushi, indie and horror films, stephen king, neosoul, hip hop,\nr&amp;b, blues, some jazz, some reggae, some classical." "novels by dostoevsky, films: the good, the bad, the ugly, lawrence\nof arabia, pan's labyrinth, in the mood for"| __truncated__ ...
#>  $ essay5: chr [1:750] "friends. . .books. . .paper to write on. . .pencils. . .music. .\n.poetry" "deserted island:<br />\n<br />\na big knife.<br />\na big pot.<br />\na big drum." "my cell phone<br />\nmy tv (don't watch it, just like hearing the voices. lol.)<br />\nmy son<br />\nmy lap top"| __truncated__ "family, friends, pets, money, the internet, food... oh, it's\nsupposed to be \"things\" only? i guess i didn't "| __truncated__ ...
#>  $ essay6: chr [1:750] "romance. . .ending capitalism and the state. . .blowing trees. .\n.sex. . . writing. . . painting. . .sex. . .e"| __truncated__ "everything. the way things are going. the intricacies of things and\nthe relative meaningless of everything.<br"| __truncated__ "moving the hell out of oakland!<br />\ngetting that first full time paycheck once i am working in my\nfield.<br"| __truncated__ "how to make my life a paradise on earth" ...
#>  $ essay7: chr [1:750] "usually a house party in the east bay.<br />\nreading leftist literature lit.<br />\nmaking art.<br />\ntalking"| __truncated__ "making music with others or making love to myself." "at home hanging with my son til his bed time. then maybe a little\ntv or internet til around one am." "having dinner in a savory restaurant and keeping in touch with\nfriends" ...
#>  $ essay8: chr [1:750] "i have a thing for hammers and sickles. (brownie points if you can\nfigure that out)<br />\n<br />\nand. . . i "| __truncated__ "this website has not done me any good." "on many occasions, i have done that cute thing people do in the\nmovies where you practice talking to someone in the mirror." "is that i don't admit private things in public" ...
#>  $ essay9: chr [1:750] "you're interested. wanna chill. wanna smoke. wanna talk politics\nand pop culture. . .wanna bash glee<br />\n<b"| __truncated__ "you're a daydreamer... and/or a nightowl.<br />\n<br />\nbut who am i to tell you what you should do? you know "| __truncated__ "you think i sound really cool and you feel like we share alot of\nthe same qualities." "you'd like to know me better and hope we will click." ...