Show token output of recipe

Returns the tokens as a list of character vector of a recipe. This function can be useful for diagnostics doing recipe construction but should not be used in final recipe steps. Note that this function will both prep() and bake() the recipe it is used on.

show_tokens(rec, var)

Arguments

rec
A recipe object

var
name of variable

Value

A list of character vectors

Examples

text_tibble <- tibble(text = c("This is words", "They are nice!"))

recipe(~ text, data = text_tibble) %>%
  step_tokenize(text) %>%
  show_tokens(text)

#> [[1]]
#> [1] "this"  "is"    "words"
#> 
#> [[2]]
#> [1] "they" "are"  "nice"
#>