Generate the basic set of text features

step_sequence_onehot creates a specification of a recipe step that will take a string and do one hot encoding for each character by position.

step_sequence_onehot(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  sequence_length = 100,
  padding = "pre",
  truncating = "pre",
  vocabulary = NULL,
  prefix = "seq1hot",
  skip = FALSE,
  id = rand_id("sequence_onehot")
)

# S3 method for step_sequence_onehot
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See recipes::selections() for more details.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

columns
A character string of variable names that will be populated (eventually) by the terms argument. This is NULL until the step is trained by recipes::prep.recipe().

sequence_length
A numeric, number of characters to keep before discarding. Defaults to 100.

padding
'pre' or 'post', pad either before or after each sequence. defaults to 'pre'.

truncating
'pre' or 'post', remove values from sequences larger than sequence_length either in the beginning or in the end of the sequence. Defaults too 'pre'.

vocabulary
A character vector, characters to be mapped to integers. Characters not in the vocabulary will be encoded as 0. Defaults to letters.

prefix
A prefix for generated column names, default to "seq1hot".

skip
A logical. Should the step be skipped when the recipe is baked by recipes::bake.recipe()? While all operations are baked when recipes::prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = FALSE.

id
A character string that is unique to this step to identify it.

x
A step_sequence_onehot object.

Source

https://papers.nips.cc/paper/5782-character-level-convolutional-networks-for-text-classification.pdf

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

The string will be capped by the sequence_length argument, strings shorter then sequence_length will be padded with empty characters. The encoding will assign a integer to each character in the vocabulary, and will encode accordingly. Characters not in the vocabulary will be encoded as 0.

See also

Examples

library(recipes)
library(modeldata)
data(okc_text)

okc_rec <- recipe(~essay0, data = okc_text) %>%
  step_tokenize(essay0) %>%
  step_tokenfilter(essay0) %>%
  step_sequence_onehot(essay0)

okc_obj <- okc_rec %>%
  prep()

bake(okc_obj, new_data = NULL)

#> # A tibble: 750 × 100
#>    seq1hot_essay0_1 seq1hot_essay0_2 seq1hot_essay0_3 seq1hot_essay0_4
#>               <dbl>            <dbl>            <dbl>            <dbl>
#>  1                0                0                0                0
#>  2                0                0                0                0
#>  3               35               98               51               84
#>  4                0                0                0                0
#>  5                0                0                0                0
#>  6               65               58               36               62
#>  7               48               84               35               65
#>  8                0                0                0                0
#>  9                0                0                0                0
#> 10                8               51                8               67
#> # … with 740 more rows, and 96 more variables: seq1hot_essay0_5 <dbl>,
#> #   seq1hot_essay0_6 <dbl>, seq1hot_essay0_7 <dbl>, seq1hot_essay0_8 <dbl>,
#> #   seq1hot_essay0_9 <dbl>, seq1hot_essay0_10 <dbl>, seq1hot_essay0_11 <dbl>,
#> #   seq1hot_essay0_12 <dbl>, seq1hot_essay0_13 <dbl>, seq1hot_essay0_14 <dbl>,
#> #   seq1hot_essay0_15 <dbl>, seq1hot_essay0_16 <dbl>, seq1hot_essay0_17 <dbl>,
#> #   seq1hot_essay0_18 <dbl>, seq1hot_essay0_19 <dbl>, seq1hot_essay0_20 <dbl>,
#> #   seq1hot_essay0_21 <dbl>, seq1hot_essay0_22 <dbl>, …

tidy(okc_rec, number = 1)

#> # A tibble: 1 × 3
#>   terms  value id            
#>   <chr>  <chr> <chr>         
#> 1 essay0 NA    tokenize_4HsIY

tidy(okc_obj, number = 1)

#> # A tibble: 1 × 3
#>   terms      value id            
#>   <quos>     <chr> <chr>         
#> 1 essay0     words tokenize_4HsIY