Generate the basic set of text features

`step_textfeature` creates a *specification* of a recipe step that will extract a number of numeric features of a text column.

step_textfeature(recipe, ..., role = "predictor", trained = FALSE,
  columns = NULL, extract_functions = count_functions,
  prefix = "textfeature", skip = FALSE, id = rand_id("textfeature"))

# S3 method for step_textfeature
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose variables. For `step_textfeature`, this indicates the variables to be encoded into a list column. See [recipes::selections()] for more details. For the `tidy` method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the recipe has been baked.

columns
A list of tibble results that define the encoding. This is `NULL` until the step is trained by [recipes::prep.recipe()].

extract_functions
A named list of feature extracting functions. default to [count_functions] from the textfeatures package. See details for more information.

prefix
A prefix for generated column names, default to "textfeature".

skip
A logical. Should the step be skipped when the recipe is baked by [recipes::bake.recipe()]? While all operations are baked when [recipes::prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it

x
A `step_textfeature` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any).

Details

This step will take a character column and returns a number of numeric columns equal to the number of functions in the list passed to the `extract_functions` argument. The default is a list of functions from the textfeatures package.

All the functions passed to `extract_functions` must take a character vector as input and return a numeric vector of the same length, otherwise an error will be thrown.

Examples

library(recipes)

data(okc_text)

okc_rec <- recipe(~ ., data = okc_text) %>%
  step_textfeature(essay0)

okc_obj <- okc_rec %>%
  prep(training = okc_text, retain = TRUE)

juice(okc_obj) %>%
  slice(1:2)

#> # A tibble: 2 x 36
#>   essay1 essay2 essay3 essay4 essay5 essay6 essay7 essay8 essay9
#>   <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct> 
#> 1 "i ju… "writ… "that… "musi… "frie… "roma… "usua… "i ha… "you'…
#> 2 eatin… "pick… "i lo… "non-… "dese… "ever… makin… this … "you'…
#> # … with 27 more variables: textfeature_essay0_n_words <int>,
#> #   textfeature_essay0_n_uq_words <int>, textfeature_essay0_n_charS <int>,
#> #   textfeature_essay0_n_uq_charS <int>, textfeature_essay0_n_digits <int>,
#> #   textfeature_essay0_n_hashtags <int>,
#> #   textfeature_essay0_n_uq_hashtags <int>,
#> #   textfeature_essay0_n_mentions <int>,
#> #   textfeature_essay0_n_uq_mentions <int>, textfeature_essay0_n_commas <int>,
#> #   textfeature_essay0_n_periods <int>, textfeature_essay0_n_exclaims <int>,
#> #   textfeature_essay0_n_extraspaces <int>, textfeature_essay0_n_caps <int>,
#> #   textfeature_essay0_n_lowers <int>, textfeature_essay0_n_urls <int>,
#> #   textfeature_essay0_n_uq_urls <int>, textfeature_essay0_n_nonasciis <int>,
#> #   textfeature_essay0_n_puncts <int>, textfeature_essay0_politeness <dbl>,
#> #   textfeature_essay0_first_person <int>,
#> #   textfeature_essay0_first_personp <int>,
#> #   textfeature_essay0_second_person <int>,
#> #   textfeature_essay0_second_personp <int>,
#> #   textfeature_essay0_third_person <int>, textfeature_essay0_to_be <int>,
#> #   textfeature_essay0_prepositions <int>

juice(okc_obj) %>%
  pull(textfeature_essay0_n_words)

#>   [1] 185  24 372  77  78 279 330 154   4 572 392 524 101  44  11 416 155  68
#>  [19]  87 115  44 115  14 107  15  73  77  83 248 302  57 249 181  49  57 106
#>  [37] 116 359  43 134 137 167  46  95  92  97 312 201 419  67  28  78  11  68
#>  [55]  33 129  31 116  38 159  43 117  28  81  86  64   8 263  32  55  92 184
#>  [73] 224  86 193  81 211 151  11  72 208 147  40  38  67 144  62  19 153 127
#>  [91]  44  15  49 216  94  30 236 197  72 114 191  26 102  44  92 232 115 116
#> [109] 164  27  70  70 195 273  64 176 130  21  91  44 126 154 120  96 122  35
#> [127]  77   3  67 160 158 100  49 143  79 195  59 243  47  74 124 106  87 156
#> [145]  87  67  60 450  44 107  74 210  13 242 176  28  81  71  45 250 729  55
#> [163] 131 102  67  66   6 137 185  83  17  33 288 625  12  37 231 143  66 156
#> [181] 157  84 123 346 164 287 232 260 135 462 102 143 368 109  94 189 217 130
#> [199]  85 146  53  71 215 152 133 167 172   5 433  37  16  84 159 138 217 202
#> [217]  46  49  46  89 150 266  66 247  44 140   7  78 109 112 114 200   4  21
#> [235] 196  43  27 288  10 172  42  34 165  95  14  15  34 134  73 654  27 306
#> [253]  66  32   8 157 228 151  77 135 155 118   3  49 295 155  97  43  34  57
#> [271]  15  50   8  43 112  51 105  39  26   7  50  84  43 160 284 333 162  69
#> [289] 100 102  82   3  29  11 266  28 120  42 227 149  49 398  65  61 114 106
#> [307] 205 146  55  40 162  95  39  64  52 167 110  79  11  36  58 575 312  28
#> [325] 102  40 133 166 441 140  24 139  10  61 109  56 139  56  35  28 457  83
#> [343]   9 212  70 251 359 227 189   5  20  61 180 190 442 122 175 130  49 137
#> [361]  30 211  90 179  19  17  83 360 100 267  77 428 378  54 313  97 541  23
#> [379]  84  83 302 236  72 329 110 461  10  89  26 159  77  31 173 109 120 249
#> [397]  43  83 299  30  36 707  81 159 181  41  75 197 161 123 259  48 101   5
#> [415]  63 223   8 142 126  49 110 213 143 226  89  62  52  40  78  53 219 284
#> [433] 193 182  41  30 113  36 220  24 571 129 275  23 131  62  56 187  98  99
#> [451] 307   6 137 149 249 131   3  11 212  31 123  95  93 239 314  79 228  74
#> [469] 135  62  68 270  70 176 125 104 165  32  99  54  52  35 105  65   3  69
#> [487]  96 422  25  74  66 110 120  95 193 435 176  13 214  85 215 129  31  85
#> [505] 114 101 140  60  17  50  29  36 105 223 210 211   9  51  25  60  14  99
#> [523]  59  59  80  98  23 683  70  38  49  22  61 119 146  97  83  48 217 147
#> [541] 187 138  62 117 221  35 335 154 199  56  78 214 289 121 353  79 153 497
#> [559] 117  56  31 100 419  94 166 128 639  85 200  65  89 244 150 255  87 185
#> [577] 320  60 167 209 204 181 122 316  44  92  26 138  78  75 276 100  47  70
#> [595]  41  89 164  10  74  27 123  25 554 223  47  99 257 184 250  20  45 325
#> [613] 423   9  41  46 136 166  99  49 185  46 200 124  89 295 322 166 129 120
#> [631] 101  97 102  78 135 287  68   7 139 234 512  14  46  58 221 213 224 499
#> [649] 189 109   7   5  37  84 138 497  97 157 103  76  74 141 133  83  73   6
#> [667]  46  27  25  45 211  30  93  34  61 279  60  69  11  23   7 532 230 104
#> [685]  38 172  76 204  12 129  49 150  59  41 103 123  28  67   9 323 219 424
#> [703]  12  45  85 333 170  79 378  92 350 220 174 168  49  94  80 114   2  23
#> [721]   6 131  72 144 168 148 164 256 295 218  57 240  39 187  76 375  20 153
#> [739] 111  40 110 118  18  61  15 104  55 247  92  45

tidy(okc_rec, number = 1)

#> # A tibble: 1 x 3
#>   terms  functions id               
#>   <chr>  <chr>     <chr>            
#> 1 essay0 <NA>      textfeature_qZ3vE

tidy(okc_obj, number = 1)

#> # A tibble: 27 x 3
#>    terms  functions     id               
#>    <chr>  <chr>         <chr>            
#>  1 essay0 n_words       textfeature_qZ3vE
#>  2 essay0 n_uq_words    textfeature_qZ3vE
#>  3 essay0 n_charS       textfeature_qZ3vE
#>  4 essay0 n_uq_charS    textfeature_qZ3vE
#>  5 essay0 n_digits      textfeature_qZ3vE
#>  6 essay0 n_hashtags    textfeature_qZ3vE
#>  7 essay0 n_uq_hashtags textfeature_qZ3vE
#>  8 essay0 n_mentions    textfeature_qZ3vE
#>  9 essay0 n_uq_mentions textfeature_qZ3vE
#> 10 essay0 n_commas      textfeature_qZ3vE
#> # … with 17 more rows

# Using custom extraction functions
nchar_round_10 <- function(x) round(nchar(x) / 10) * 10

recipe(~ ., data = okc_text) %>%
  step_textfeature(essay0,
                   extract_functions = list(nchar10 = nchar_round_10)) %>%
  prep(training = okc_text) %>%
  juice()

#> # A tibble: 750 x 10
#>    essay1 essay2 essay3 essay4 essay5 essay6 essay7 essay8 essay9
#>    <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct> 
#>  1 "i ju… "writ… "that… "musi… "frie… "roma… "usua… "i ha… "you'…
#>  2 eatin… "pick… "i lo… "non-… "dese… "ever… makin… this … "you'…
#>  3 "rais… "proc… "my s… "sush… "my c… "movi… "at h… "on m… "you …
#>  4 "i ju… "i've… my fa… "nove… "fami… how t… "havi… is th… you'd…
#>  5 i'm a… being… "that… "book… "- fr… "the … proba… "uhh.… you s…
#>  6 "i'm … "i'm … "the … "book… "guit… "a li… "hang… "i'm … "if y…
#>  7 "well… "well… "eith… "i do… "1) m… "the … "out … "i ow… "you …
#>  8 "i wo… "well… dimpl… "book… "1-la… "sex<… "depe… i lik… "you …
#>  9 photo… eatin… my gl… donni… sushi… love.  drink… hm...  .     
#> 10 "occu… "livi… long … "film… yuzu,… "menu… "ofte… but i… "you …
#> # … with 740 more rows, and 1 more variable: textfeature_essay0_nchar10 <dbl>