Term frequency of tokens

`step_texthash` creates a *specification* of a recipe step that will convert a list of tokens into multiple variables using the hashing trick.

step_texthash(recipe, ..., role = NA, trained = FALSE,
  columns = NULL, signed = TRUE, num_terms = 1024, prefix = "hash",
  skip = FALSE, id = rand_id("texthash"))

# S3 method for step_texthash
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose variables. For `step_texthash`, this indicates the variables to be encoded into a list column. See [recipes::selections()] for more details. For the `tidy` method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the recipe has been baked.

columns
A list of tibble results that define the encoding. This is `NULL` until the step is trained by [recipes::prep.recipe()].

signed
A logical, indicating whether to use a signed hash-function to reduce collisions when hashing. Defaults to TRUE.

num_terms
An integer, the number of variables to output. Defaults to 1024.

prefix
A character string that will be the prefix to the resulting new variables. See notes below.

skip
A logical. Should the step be skipped when the recipe is baked by [recipes::bake.recipe()]? While all operations are baked when [recipes::prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it.

x
A `step_texthash` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any).

Details

Feature hashing, or the hashing trick, is a transformation of a text variable into a new set of numerical variables. This is done by applying a hashing function over the tokens and using the hash values as feature indices. This allows for a low memory representation of the text. This implementation is done using the MurmurHash3 method.

The argument `num_terms` controls the number of indices that the hashing function will map to. This is the tuning parameter for this transformation. Since the hashing function can map two different tokens to the same index, will a higher value of `num_terms` result in a lower chance of collision.

The new components will have names that begin with `prefix`, then the name of the variable, followed by the tokens all seperated by `-`. The variable names are padded with zeros. For example, if `num_terms < 10`, their names will be `hash1` - `hash9`. If `num_terms = 101`, their names will be `hash001` - `hash101`.

References

Kilian Weinberger; Anirban Dasgupta; John Langford; Alex Smola; Josh Attenberg (2009).

See also

Examples

library(recipes)

data(okc_text)

okc_rec <- recipe(~ ., data = okc_text) %>%
  step_tokenize(essay0) %>%
  step_tokenfilter(essay0, max_tokens = 10) %>%
  step_texthash(essay0)

okc_obj <- okc_rec %>%
  prep(training = okc_text, retain = TRUE)

bake(okc_obj, okc_text)

#> # A tibble: 750 x 1,033
#>    essay1 essay2 essay3 essay4 essay5 essay6 essay7 essay8 essay9
#>    <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct>  <fct> 
#>  1 "i ju… "writ… "that… "musi… "frie… "roma… "usua… "i ha… "you'…
#>  2 eatin… "pick… "i lo… "non-… "dese… "ever… makin… this … "you'…
#>  3 "rais… "proc… "my s… "sush… "my c… "movi… "at h… "on m… "you …
#>  4 "i ju… "i've… my fa… "nove… "fami… how t… "havi… is th… you'd…
#>  5 i'm a… being… "that… "book… "- fr… "the … proba… "uhh.… you s…
#>  6 "i'm … "i'm … "the … "book… "guit… "a li… "hang… "i'm … "if y…
#>  7 "well… "well… "eith… "i do… "1) m… "the … "out … "i ow… "you …
#>  8 "i wo… "well… dimpl… "book… "1-la… "sex<… "depe… i lik… "you …
#>  9 photo… eatin… my gl… donni… sushi… love.  drink… hm...  .     
#> 10 "occu… "livi… long … "film… yuzu,… "menu… "ofte… but i… "you …
#> # ... with 740 more rows, and 1,024 more variables: essay0_hash0001 <dbl>,
#> #   essay0_hash0002 <dbl>, essay0_hash0003 <dbl>, essay0_hash0004 <dbl>,
#> #   essay0_hash0005 <dbl>, essay0_hash0006 <dbl>, essay0_hash0007 <dbl>,
#> #   essay0_hash0008 <dbl>, essay0_hash0009 <dbl>, essay0_hash0010 <dbl>,
#> #   essay0_hash0011 <dbl>, essay0_hash0012 <dbl>, essay0_hash0013 <dbl>,
#> #   essay0_hash0014 <dbl>, essay0_hash0015 <dbl>, essay0_hash0016 <dbl>,
#> #   essay0_hash0017 <dbl>, essay0_hash0018 <dbl>, essay0_hash0019 <dbl>,
#> #   essay0_hash0020 <dbl>, essay0_hash0021 <dbl>, essay0_hash0022 <dbl>,
#> #   essay0_hash0023 <dbl>, essay0_hash0024 <dbl>, essay0_hash0025 <dbl>,
#> #   essay0_hash0026 <dbl>, essay0_hash0027 <dbl>, essay0_hash0028 <dbl>,
#> #   essay0_hash0029 <dbl>, essay0_hash0030 <dbl>, essay0_hash0031 <dbl>,
#> #   essay0_hash0032 <dbl>, essay0_hash0033 <dbl>, essay0_hash0034 <dbl>,
#> #   essay0_hash0035 <dbl>, essay0_hash0036 <dbl>, essay0_hash0037 <dbl>,
#> #   essay0_hash0038 <dbl>, essay0_hash0039 <dbl>, essay0_hash0040 <dbl>,
#> #   essay0_hash0041 <dbl>, essay0_hash0042 <dbl>, essay0_hash0043 <dbl>,
#> #   essay0_hash0044 <dbl>, essay0_hash0045 <dbl>, essay0_hash0046 <dbl>,
#> #   essay0_hash0047 <dbl>, essay0_hash0048 <dbl>, essay0_hash0049 <dbl>,
#> #   essay0_hash0050 <dbl>, essay0_hash0051 <dbl>, essay0_hash0052 <dbl>,
#> #   essay0_hash0053 <dbl>, essay0_hash0054 <dbl>, essay0_hash0055 <dbl>,
#> #   essay0_hash0056 <dbl>, essay0_hash0057 <dbl>, essay0_hash0058 <dbl>,
#> #   essay0_hash0059 <dbl>, essay0_hash0060 <dbl>, essay0_hash0061 <dbl>,
#> #   essay0_hash0062 <dbl>, essay0_hash0063 <dbl>, essay0_hash0064 <dbl>,
#> #   essay0_hash0065 <dbl>, essay0_hash0066 <dbl>, essay0_hash0067 <dbl>,
#> #   essay0_hash0068 <dbl>, essay0_hash0069 <dbl>, essay0_hash0070 <dbl>,
#> #   essay0_hash0071 <dbl>, essay0_hash0072 <dbl>, essay0_hash0073 <dbl>,
#> #   essay0_hash0074 <dbl>, essay0_hash0075 <dbl>, essay0_hash0076 <dbl>,
#> #   essay0_hash0077 <dbl>, essay0_hash0078 <dbl>, essay0_hash0079 <dbl>,
#> #   essay0_hash0080 <dbl>, essay0_hash0081 <dbl>, essay0_hash0082 <dbl>,
#> #   essay0_hash0083 <dbl>, essay0_hash0084 <dbl>, essay0_hash0085 <dbl>,
#> #   essay0_hash0086 <dbl>, essay0_hash0087 <dbl>, essay0_hash0088 <dbl>,
#> #   essay0_hash0089 <dbl>, essay0_hash0090 <dbl>, essay0_hash0091 <dbl>,
#> #   essay0_hash0092 <dbl>, essay0_hash0093 <dbl>, essay0_hash0094 <dbl>,
#> #   essay0_hash0095 <dbl>, essay0_hash0096 <dbl>, essay0_hash0097 <dbl>,
#> #   essay0_hash0098 <dbl>, essay0_hash0099 <dbl>, essay0_hash0100 <dbl>, …

tidy(okc_rec, number = 2)

#> # A tibble: 1 x 3
#>   terms  value id               
#>   <chr>  <int> <chr>            
#> 1 essay0    NA tokenfilter_AR6ln

tidy(okc_obj, number = 2)

#> # A tibble: 1 x 3
#>   terms          value     id               
#>   <S3: quosures> <list>    <chr>            
#> 1 ~essay0        <int [1]> tokenfilter_AR6ln