Feature Hashing of Tokens

step_texthash creates a specification of a recipe step that will convert a token variable into multiple numeric variables using the hashing trick.

Usage

step_texthash(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  signed = TRUE,
  num_terms = 1024L,
  prefix = "texthash",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("texthash")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose which variables are affected by the step. See recipes::selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

columns

A character string of variable names that will be populated (eventually) by the terms argument. This is NULL until the step is trained by recipes::prep.recipe().

signed

A logical, indicating whether to use a signed hash-function to reduce collisions when hashing. Defaults to TRUE.

num_terms

An integer, the number of variables to output. Defaults to 1024.

prefix

A character string that will be the prefix to the resulting new variables. See notes below.

keep_original_cols

A logical to keep the original variables in the output. Defaults to FALSE.

skip

A logical. Should the step be skipped when the recipe is baked by recipes::bake.recipe()? While all operations are baked when recipes::prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = FALSE.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

Feature hashing, or the hashing trick, is a transformation of a text variable into a new set of numerical variables. This is done by applying a hashing function over the tokens and using the hash values as feature indices. This allows for a low memory representation of the text. This implementation is done using the MurmurHash3 method.

The argument num_terms controls the number of indices that the hashing function will map to. This is the tuning parameter for this transformation. Since the hashing function can map two different tokens to the same index, will a higher value of num_terms result in a lower chance of collision.

The new components will have names that begin with prefix, then the name of the variable, followed by the tokens all separated by -. The variable names are padded with zeros. For example if prefix = "hash", and if num_terms < 10, their names will be hash1 - hash9. If num_terms = 101, their names will be hash001 - hash101.

Tidying

When you tidy() this step, a tibble with columns terms (the selectors or variables selected) and value (number of terms).

Tuning Parameters

This step has 2 tuning parameters:

signed: Signed Hash Value (type: logical, default: TRUE)

num_terms: # Hash Features (type: integer, default: 1024)

Case weights

The underlying operation does not allow for case weights.

References

Kilian Weinberger; Anirban Dasgupta; John Langford; Alex Smola; Josh Attenberg (2009).

See also

Examples

library(recipes)
library(modeldata)
data(tate_text)

tate_rec <- recipe(~., data = tate_text) %>%
  step_tokenize(medium) %>%
  step_tokenfilter(medium, max_tokens = 10) %>%
  step_texthash(medium)

tate_obj <- tate_rec %>%
  prep()

bake(tate_obj, tate_text)
#> # A tibble: 4,284 × 1,028
#>    texthash_medium_0001 texthash_medium_0002 texthash_medium_0003
#>                   <int>                <int>                <int>
#>  1                    0                    0                    0
#>  2                    0                    0                    0
#>  3                    0                    0                    0
#>  4                    0                    0                    0
#>  5                    0                    0                    0
#>  6                    0                    0                    0
#>  7                    0                    0                    0
#>  8                    0                    0                    0
#>  9                    0                    0                    0
#> 10                    0                    0                    0
#> # ℹ 4,274 more rows
#> # ℹ 1,025 more variables: texthash_medium_0004 <int>,
#> #   texthash_medium_0005 <int>, texthash_medium_0006 <int>,
#> #   texthash_medium_0007 <int>, texthash_medium_0008 <int>,
#> #   texthash_medium_0009 <int>, texthash_medium_0010 <int>,
#> #   texthash_medium_0011 <int>, texthash_medium_0012 <int>,
#> #   texthash_medium_0013 <int>, texthash_medium_0014 <int>, …

tidy(tate_rec, number = 3)
#> # A tibble: 1 × 4
#>   terms  value length id            
#>   <chr>  <lgl>  <int> <chr>         
#> 1 medium NA        NA texthash_3yd68
tidy(tate_obj, number = 3)
#> # A tibble: 1 × 4
#>   terms  value length id            
#>   <chr>  <lgl>  <int> <chr>         
#> 1 medium TRUE    1024 texthash_3yd68