Term frequency of tokens

step_tf creates a specification of a recipe step that will convert a tokenlist into multiple variables containing the token counts.

step_tf(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  weight_scheme = "raw count",
  weight = 0.5,
  vocabulary = NULL,
  res = NULL,
  prefix = "tf",
  skip = FALSE,
  id = rand_id("tf")
)

# S3 method for step_tf
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose variables. For step_tf, this indicates the variables to be encoded into a tokenlist. See recipes::selections() for more details. For the tidy method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the recipe has been baked.

columns
A list of tibble results that define the encoding. This is NULL until the step is trained by recipes::prep.recipe().

weight_scheme
A character determining the weighting scheme for the term frequency calculations. Must be one of "binary", "raw count", "term frequency", "log normalization" or "double normalization". Defaults to "raw count".

weight
A numeric weight used if weight_scheme is set to "double normalization". Defaults to 0.5.

vocabulary
A character vector of strings to be considered.

res
The words that will be used to calculate the term frequency will be stored here once this preprocessing step has be trained by prep.recipe().

prefix
A character string that will be the prefix to the resulting new variables. See notes below

skip
A logical. Should the step be skipped when the recipe is baked by recipes::bake.recipe()? While all operations are baked when recipes::prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it.

x
A step_tf object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

It is strongly advised to use step_tokenfilter before using step_tf to limit the number of variables created, otherwise you might run into memory issues. A good strategy is to start with a low token count and go up according to how much RAM you want to use.

Term frequency is a weight of how many times each token appear in each observation. There are different ways to calculate the weight and this step can do it in a couple of ways. Setting the argument weight_scheme to "binary" will result in a set of binary variables denoting if a token is present in the observation. "raw count" will count the times a token is present in the observation. "term frequency" will divide the count with the total number of words in the document to limit the effect of the document length as longer documents tends to have the word present more times but not necessarily at a higher percentage. "log normalization" takes the log of 1 plus the count, adding 1 is done to avoid taking log of 0. Finally "double normalization" is the raw frequency divided by the raw frequency of the most occurring term in the document. This is then multiplied by weight and weight is added to the result. This is again done to prevent a bias towards longer documents.

The new components will have names that begin with prefix, then the name of the variable, followed by the tokens all separated by -. The new variables will be created alphabetically according to token.

See also

Examples

# \donttest{
library(recipes)
library(modeldata)
data(okc_text)

okc_rec <- recipe(~., data = okc_text) %>%
  step_tokenize(essay0) %>%
  step_tf(essay0)

okc_obj <- okc_rec %>%
  prep()

bake(okc_obj, okc_text)

#> # A tibble: 750 x 9,265
#>    essay1   essay2   essay3   essay4   essay5   essay6   essay7  essay8  essay9 
#>    <fct>    <fct>    <fct>    <fct>    <fct>    <fct>    <fct>   <fct>   <fct>  
#>  1 "i just… "writin… "that i… "music:… "friend… "romanc… "usual… "i hav… "you'r…
#>  2 "eating… "pickin… "i look… "non-fi… "desert… "everyt… "makin… "this … "you'r…
#>  3 "raisin… "procra… "my smi… "sushi,… "my cel… "moving… "at ho… "on ma… "you t…
#>  4 "i just… "i've b… "my fas… "novels… "family… "how to… "havin… "is th… "you'd…
#>  5 "i'm a … "being … "that w… "books … "- frie… "the un… "proba… "uhh..… "you s…
#>  6 "i'm cu… "i'm re… "the wa… "books … "guitar… "a litt… "hangi… "i'm p… "if yo…
#>  7 "well, … "well, … "either… "i don'… "1) my … "the en… "out w… "i own… "you a…
#>  8 "i work… "welll.… "dimple… "book-t… "1-laug… "sex<br… "depen… "i lik… "you a…
#>  9 "photog… "eating… "my gla… "donnie… "sushi.… "love."  "drink… "hm..." "."    
#> 10 "occupy… "living… "long h… "films<… "yuzu, … "menus,… "often… "but i… "you a…
#> # … with 740 more rows, and 9,256 more variables: tf_essay0______ <dbl>,
#> #   tf_essay0_______ <dbl>, tf_essay0________________________ <dbl>,
#> #   tf_essay0_____coming <dbl>, tf_essay0__blank <dbl>,
#> #   tf_essay0__updates_ <dbl>, tf_essay0_0 <dbl>, tf_essay0_01 <dbl>,
#> #   tf_essay0_0aare <dbl>, tf_essay0_0abilly <dbl>,
#> #   tf_essay0_0aboondocks <dbl>, tf_essay0_0abrothers <dbl>,
#> #   tf_essay0_0aconfidential <dbl>, tf_essay0_0aconversation <dbl>,
#> #   tf_essay0_0adebates <dbl>, tf_essay0_0afly <dbl>,
#> #   tf_essay0_0afriends <dbl>, tf_essay0_0agiants <dbl>, tf_essay0_0ahop <dbl>,
#> #   tf_essay0_0ahunters <dbl>, tf_essay0_0aking <dbl>,
#> #   tf_essay0_0amovies <dbl>, tf_essay0_0amusic <dbl>,
#> #   tf_essay0_0aparties <dbl>, tf_essay0_0arailroading <dbl>,
#> #   tf_essay0_0ashows <dbl>, tf_essay0_0atrips <dbl>,
#> #   tf_essay0_0aweapons <dbl>, tf_essay0_1 <dbl>, tf_essay0_10 <dbl>,
#> #   tf_essay0_10,000 <dbl>, tf_essay0_100 <dbl>, tf_essay0_1000 <dbl>,
#> #   tf_essay0_105 <dbl>, tf_essay0_11 <dbl>, tf_essay0_110 <dbl>,
#> #   tf_essay0_1193 <dbl>, tf_essay0_12 <dbl>, tf_essay0_125 <dbl>,
#> #   tf_essay0_12s <dbl>, tf_essay0_13 <dbl>, tf_essay0_1337 <dbl>,
#> #   tf_essay0_14 <dbl>, tf_essay0_1400 <dbl>, tf_essay0_15 <dbl>,
#> #   tf_essay0_150 <dbl>, tf_essay0_16 <dbl>, tf_essay0_16th <dbl>,
#> #   tf_essay0_17 <dbl>, tf_essay0_18 <dbl>, tf_essay0_180 <dbl>,
#> #   tf_essay0_1886866717 <dbl>, tf_essay0_19 <dbl>, tf_essay0_1904 <dbl>,
#> #   tf_essay0_1964 <dbl>, tf_essay0_1966 <dbl>, tf_essay0_1982 <dbl>,
#> #   tf_essay0_1988 <dbl>, tf_essay0_1991 <dbl>, tf_essay0_1992 <dbl>,
#> #   tf_essay0_1996 <dbl>, tf_essay0_1998 <dbl>, tf_essay0_1st <dbl>,
#> #   tf_essay0_2 <dbl>, tf_essay0_20 <dbl>, tf_essay0_200 <dbl>,
#> #   tf_essay0_2000s <dbl>, tf_essay0_2001 <dbl>, tf_essay0_2005 <dbl>,
#> #   tf_essay0_2007 <dbl>, tf_essay0_2008 <dbl>, tf_essay0_2009 <dbl>,
#> #   tf_essay0_2010 <dbl>, tf_essay0_2011 <dbl>, tf_essay0_2012 <dbl>,
#> #   tf_essay0_202 <dbl>, tf_essay0_2021 <dbl>, tf_essay0_20s <dbl>,
#> #   tf_essay0_20snot <dbl>, tf_essay0_20th <dbl>, tf_essay0_21 <dbl>,
#> #   tf_essay0_22 <dbl>, tf_essay0_23 <dbl>, tf_essay0_23yo <dbl>,
#> #   tf_essay0_24 <dbl>, tf_essay0_245lb <dbl>, tf_essay0_25 <dbl>,
#> #   tf_essay0_250 <dbl>, tf_essay0_26 <dbl>, tf_essay0_27 <dbl>,
#> #   tf_essay0_27ish <dbl>, tf_essay0_27s <dbl>, tf_essay0_28 <dbl>,
#> #   tf_essay0_28th <dbl>, tf_essay0_29 <dbl>, tf_essay0_2cedd <dbl>,
#> #   tf_essay0_2cehksxolna <dbl>, tf_essay0_2fbowling <dbl>,
#> #   tf_essay0_2fdarts <dbl>, tf_essay0_2fodd <dbl>, …

tidy(okc_rec, number = 2)

#> # A tibble: 1 x 3
#>   terms  value id      
#>   <chr>  <chr> <chr>   
#> 1 essay0 NA    tf_IlOjL

tidy(okc_obj, number = 2)

#> # A tibble: 1 x 3
#>   terms      value     id      
#>   <quos>     <chr>     <chr>   
#> 1 essay0     raw count tf_IlOjL

# }