Term frequency of tokens

step_tf creates a specification of a recipe step that will convert a tokenlist into multiple variables containing the token counts.

step_tf(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  weight_scheme = "raw count",
  weight = 0.5,
  vocabulary = NULL,
  res = NULL,
  prefix = "tf",
  skip = FALSE,
  id = rand_id("tf")
)

# S3 method for step_tf
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See recipes::selections() for more details.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

columns
A character string of variable names that will be populated (eventually) by the terms argument. This is NULL until the step is trained by recipes::prep.recipe().

weight_scheme
A character determining the weighting scheme for the term frequency calculations. Must be one of "binary", "raw count", "term frequency", "log normalization" or "double normalization". Defaults to "raw count".

weight
A numeric weight used if weight_scheme is set to "double normalization". Defaults to 0.5.

vocabulary
A character vector of strings to be considered.

res
The words that will be used to calculate the term frequency will be stored here once this preprocessing step has be trained by prep.recipe().

prefix
A character string that will be the prefix to the resulting new variables. See notes below.

skip
A logical. Should the step be skipped when the recipe is baked by recipes::bake.recipe()? While all operations are baked when recipes::prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = FALSE.

id
A character string that is unique to this step to identify it.

x
A step_tf object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

It is strongly advised to use step_tokenfilter before using step_tf to limit the number of variables created, otherwise you might run into memory issues. A good strategy is to start with a low token count and go up according to how much RAM you want to use.

Term frequency is a weight of how many times each token appear in each observation. There are different ways to calculate the weight and this step can do it in a couple of ways. Setting the argument weight_scheme to "binary" will result in a set of binary variables denoting if a token is present in the observation. "raw count" will count the times a token is present in the observation. "term frequency" will divide the count with the total number of words in the document to limit the effect of the document length as longer documents tends to have the word present more times but not necessarily at a higher percentage. "log normalization" takes the log of 1 plus the count, adding 1 is done to avoid taking log of 0. Finally "double normalization" is the raw frequency divided by the raw frequency of the most occurring term in the document. This is then multiplied by weight and weight is added to the result. This is again done to prevent a bias towards longer documents.

The new components will have names that begin with prefix, then the name of the variable, followed by the tokens all separated by -. The variable names are padded with zeros. For example if prefix = "hash", and if num_terms < 10, their names will be hash1 - hash9. If num_terms = 101, their names will be hash001 - hash101.

See also

Examples

# \donttest{
library(recipes)
library(modeldata)
data(tate_text)

tate_rec <- recipe(~., data = tate_text) %>%
  step_tokenize(medium) %>%
  step_tf(medium)

tate_obj <- tate_rec %>%
  prep()

bake(tate_obj, tate_text)

#> # A tibble: 4,284 × 956
#>        id artist title  year tf_medium_1 tf_medium_10 tf_medium_100 tf_medium_11
#>     <dbl> <fct>  <fct> <dbl>       <dbl>        <dbl>         <dbl>        <dbl>
#>  1  21926 Absal… Prop…  1990           0            0             0            0
#>  2  20472 Auerb… Mich…  1990           0            0             0            0
#>  3  20474 Auerb… Geof…  1990           0            0             0            0
#>  4  20473 Auerb… Jake   1990           0            0             0            0
#>  5  20513 Auerb… To t…  1990           0            0             0            0
#>  6  21389 Ayres… Phaë…  1990           0            0             0            0
#>  7 121187 Barlo… Unti…  1990           0            0             0            0
#>  8  19455 Basel… Gree…  1990           0            0             0            0
#>  9  20938 Beatt… Pres…  1990           0            0             0            0
#> 10 105941 Beuys… Jose…  1990           0            0             0            0
#> # … with 4,274 more rows, and 948 more variables: tf_medium_12 <dbl>,
#> #   tf_medium_13 <dbl>, tf_medium_133 <dbl>, tf_medium_14 <dbl>,
#> #   tf_medium_15 <dbl>, tf_medium_151 <dbl>, tf_medium_16 <dbl>,
#> #   tf_medium_160 <dbl>, tf_medium_16mm <dbl>, tf_medium_18 <dbl>,
#> #   tf_medium_19 <dbl>, tf_medium_2 <dbl>, tf_medium_20 <dbl>,
#> #   tf_medium_2000 <dbl>, tf_medium_201 <dbl>, tf_medium_21 <dbl>,
#> #   tf_medium_22 <dbl>, tf_medium_220 <dbl>, tf_medium_23 <dbl>, …

tidy(tate_rec, number = 2)

#> # A tibble: 1 × 3
#>   terms  value id      
#>   <chr>  <chr> <chr>   
#> 1 medium NA    tf_ubgp4

tidy(tate_obj, number = 2)

#> # A tibble: 1 × 3
#>   terms      value     id      
#>   <quos>     <chr>     <chr>   
#> 1 medium     raw count tf_ubgp4

# }