Term frequency-inverse document frequency of tokens

step_tfidf creates a specification of a recipe step that will convert a tokenlist into multiple variables containing the term frequency-inverse document frequency of tokens.

step_tfidf(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  vocabulary = NULL,
  res = NULL,
  smooth_idf = TRUE,
  norm = "l1",
  sublinear_tf = FALSE,
  prefix = "tfidf",
  skip = FALSE,
  id = rand_id("tfidf")
)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See recipes::selections() for more details.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

columns
A character string of variable names that will be populated (eventually) by the terms argument. This is NULL until the step is trained by recipes::prep.recipe().

vocabulary
A character vector of strings to be considered.

res
The words that will be used to calculate the term frequency will be stored here once this preprocessing step has be trained by prep.recipe().

smooth_idf
TRUE smooth IDF weights by adding one to document frequencies, as if an extra document was seen containing every term in the collection exactly once. This prevents division by zero.

norm
A character, defines the type of normalization to apply to term vectors. "l1" by default, i.e., scale by the number of words in the document. Must be one of c("l1", "l2", "none").

sublinear_tf
A logical, apply sublinear term-frequency scaling, i.e., replace the term frequency with 1 + log(TF). Defaults to FALSE.

prefix
A character string that will be the prefix to the resulting new variables. See notes below.

skip
A logical. Should the step be skipped when the recipe is baked by recipes::bake.recipe()? While all operations are baked when recipes::prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = FALSE.

id
A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

It is strongly advised to use step_tokenfilter before using step_tfidf to limit the number of variables created; otherwise you may run into memory issues. A good strategy is to start with a low token count and increase depending on how much RAM you want to use.

Term frequency-inverse document frequency is the product of two statistics: the term frequency (TF) and the inverse document frequency (IDF).

Term frequency measures how many times each token appears in each observation.

Inverse document frequency is a measure of how informative a word is, e.g., how common or rare the word is across all the observations. If a word appears in all the observations it might not give that much insight, but if it only appears in some it might help differentiate between observations.

The IDF is defined as follows: idf = log(1 + (# documents in the corpus) / (# documents where the term appears))

The new components will have names that begin with prefix, then the name of the variable, followed by the tokens all separated by -. The variable names are padded with zeros. For example if prefix = "hash", and if num_terms < 10, their names will be hash1 - hash9. If num_terms = 101, their names will be hash001 - hash101.

See also

Examples

# \donttest{
library(recipes)
library(modeldata)
data(tate_text)

tate_rec <- recipe(~., data = tate_text) %>%
  step_tokenize(medium) %>%
  step_tfidf(medium)

tate_obj <- tate_rec %>%
  prep()

bake(tate_obj, tate_text)

#> # A tibble: 4,284 × 956
#>        id artist             title  year tfidf_medium_1 tfidf_medium_10 tfidf_medium_100
#>     <dbl> <fct>              <fct> <dbl>          <dbl>           <dbl>            <dbl>
#>  1  21926 Absalon            Prop…  1990              0               0                0
#>  2  20472 Auerbach, Frank    Mich…  1990              0               0                0
#>  3  20474 Auerbach, Frank    Geof…  1990              0               0                0
#>  4  20473 Auerbach, Frank    Jake   1990              0               0                0
#>  5  20513 Auerbach, Frank    To t…  1990              0               0                0
#>  6  21389 Ayres, OBE Gillian Phaë…  1990              0               0                0
#>  7 121187 Barlow, Phyllida   Unti…  1990              0               0                0
#>  8  19455 Baselitz, Georg    Gree…  1990              0               0                0
#>  9  20938 Beattie, Basil     Pres…  1990              0               0                0
#> 10 105941 Beuys, Joseph      Jose…  1990              0               0                0
#> # … with 4,274 more rows, and 949 more variables: tfidf_medium_11 <dbl>,
#> #   tfidf_medium_12 <dbl>, tfidf_medium_13 <dbl>, tfidf_medium_133 <dbl>,
#> #   tfidf_medium_14 <dbl>, tfidf_medium_15 <dbl>, tfidf_medium_151 <dbl>,
#> #   tfidf_medium_16 <dbl>, tfidf_medium_160 <dbl>, tfidf_medium_16mm <dbl>,
#> #   tfidf_medium_18 <dbl>, tfidf_medium_19 <dbl>, tfidf_medium_2 <dbl>,
#> #   tfidf_medium_20 <dbl>, tfidf_medium_2000 <dbl>, tfidf_medium_201 <dbl>,
#> #   tfidf_medium_21 <dbl>, tfidf_medium_22 <dbl>, tfidf_medium_220 <dbl>, …

tidy(tate_rec, number = 2)

#> # A tibble: 1 × 2
#>   terms  id         
#>   <chr>  <chr>      
#> 1 medium tfidf_BSJsJ

tidy(tate_obj, number = 2)

#> # A tibble: 1 × 2
#>   terms      id         
#>   <quos>     <chr>      
#> 1 medium     tfidf_BSJsJ

# }