Generate the basic set of text features

`step_tokenmerge` creates a *specification* of a recipe step that will take multiple list-columns of tokens and combine them into one list-column.

step_tokenmerge(recipe, ..., role = "predictor", trained = FALSE,
  columns = NULL, prefix = "tokenmerge", skip = FALSE,
  id = rand_id("tokenmerge"))

# S3 method for step_tokenmerge
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose variables. For `step_tokenmerge`, this indicates the variables to be encoded into a list column. See [recipes::selections()] for more details. For the `tidy` method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the recipe has been baked.

columns
A list of tibble results that define the encoding. This is `NULL` until the step is trained by [recipes::prep.recipe()].

prefix
A prefix for generated column names, default to "tokenmerge".

skip
A logical. Should the step be skipped when the recipe is baked by [recipes::bake.recipe()]? While all operations are baked when [recipes::prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it

x
A `step_tokenmerge` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any).

Examples

library(recipes)

data(okc_text)

okc_rec <- recipe(~ ., data = okc_text) %>%
  step_tokenize(essay0, essay1) %>%
  step_tokenmerge(essay0, essay1)

okc_obj <- okc_rec %>%
  prep(training = okc_text, retain = TRUE)

juice(okc_obj)

#> # A tibble: 750 x 9
#>    essay2   essay3   essay4   essay5  essay6  essay7  essay8  essay9  tokenmerge
#>    <fct>    <fct>    <fct>    <fct>   <fct>   <fct>   <fct>   <fct>   <list>    
#>  1 "writin… "that i… "music:… "frien… "roman… "usual… "i hav… "you'r… <chr [240…
#>  2 "pickin… "i look… "non-fi… "deser… "every… making… this w… "you'r… <chr [28]>
#>  3 "procra… "my smi… "sushi,… "my ce… "movin… "at ho… "on ma… "you t… <chr [385…
#>  4 "i've b… my fash… "novels… "famil… how to… "havin… is tha… you'd … <chr [104…
#>  5 being h… "that w… "books … "- fri… "the u… probab… "uhh..… you sh… <chr [87]>
#>  6 "i'm re… "the wa… "books … "guita… "a lit… "hangi… "i'm p… "if yo… <chr [352…
#>  7 "well, … "either… "i don'… "1) my… "the e… "out w… "i own… "you a… <chr [374…
#>  8 "welll.… dimples… "book-t… "1-lau… "sex<b… "depen… i like… "you a… <chr [228…
#>  9 eating … my glas… donnie … sushi.… love.   drinki… hm...   .       <chr [5]> 
#> 10 "living… long ha… "films<… yuzu, … "menus… "often… but it… "you a… <chr [610…
#> # … with 740 more rows

tidy(okc_rec, number = 1)

#> # A tibble: 2 x 3
#>   terms  value id            
#>   <chr>  <chr> <chr>         
#> 1 essay0 <NA>  tokenize_LGGmr
#> 2 essay1 <NA>  tokenize_LGGmr

tidy(okc_obj, number = 1)

#> # A tibble: 2 x 3
#>   terms      value id            
#>   <quos>     <chr> <chr>         
#> 1 essay0     words tokenize_LGGmr
#> 2 essay1     words tokenize_LGGmr