Create tokenlist object

A tokenlist object is a thin wrapper around a list of character vectors, with a few attributes.

tokenlist(tokens = list(), lemma = NULL, pos = NULL)

Arguments

tokens
List of character vectors

lemma
List of character vectors, must be same size and shape as x.

pos
List of character vectors, must be same size and shape as x.

Value

a tokenlist object.

Examples

abc <- list(letters, LETTERS)
tokenlist(abc)

#> <textrecipes_tokenlist[2]>
#> [1] [26 tokens] [26 tokens]
#> # Unique Tokens: 52

unclass(tokenlist(abc))

#> $tokens
#> $tokens[[1]]
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z"
#> 
#> $tokens[[2]]
#>  [1] "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L" "M" "N" "O" "P" "Q" "R" "S"
#> [20] "T" "U" "V" "W" "X" "Y" "Z"
#> 
#> 
#> attr(,"unique_tokens")
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z" "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L"
#> [39] "M" "N" "O" "P" "Q" "R" "S" "T" "U" "V" "W" "X" "Y" "Z"

tibble(text = tokenlist(abc))

#> # A tibble: 2 × 1
#>          text
#>     <tknlist>
#> 1 [26 tokens]
#> 2 [26 tokens]

library(tokenizers)
library(modeldata)
data(tate_text)
tokens <- tokenize_words(as.character(tate_text$medium))

tokenlist(tokens)

#> <textrecipes_tokenlist[4284]>
#>    [1] [8 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [4 tokens] 
#>    [7] [4 tokens]  [3 tokens]  [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>   [13] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [7 tokens] 
#>   [19] [10 tokens] [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>   [25] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens] 
#>   [31] [8 tokens]  [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>   [37] [3 tokens]  [3 tokens]  [3 tokens]  [1 tokens]  [5 tokens]  [4 tokens] 
#>   [43] [1 tokens]  [10 tokens] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>   [49] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens] 
#>   [55] [10 tokens] [5 tokens]  [6 tokens]  [7 tokens]  [4 tokens]  [4 tokens] 
#>   [61] [4 tokens]  [4 tokens]  [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>   [67] [3 tokens]  [4 tokens]  [3 tokens]  [6 tokens]  [3 tokens]  [3 tokens] 
#>   [73] [3 tokens]  [5 tokens]  [11 tokens] [3 tokens]  [3 tokens]  [4 tokens] 
#>   [79] [18 tokens] [3 tokens]  [8 tokens]  [6 tokens]  [10 tokens] [13 tokens]
#>   [85] [5 tokens]  [3 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>   [91] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens] 
#>   [97] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [3 tokens] 
#>  [103] [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [5 tokens] 
#>  [109] [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [115] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [121] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [127] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [133] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [139] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [145] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [151] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [157] [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [3 tokens]  [6 tokens] 
#>  [163] [6 tokens]  [1 tokens]  [5 tokens]  [8 tokens]  [1 tokens]  [4 tokens] 
#>  [169] [6 tokens]  [6 tokens]  [6 tokens]  [4 tokens]  [8 tokens]  [5 tokens] 
#>  [175] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [7 tokens]  [3 tokens] 
#>  [181] [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [3 tokens]  [3 tokens] 
#>  [187] [6 tokens]  [4 tokens]  [8 tokens]  [11 tokens] [4 tokens]  [11 tokens]
#>  [193] [3 tokens]  [3 tokens]  [4 tokens]  [3 tokens]  [6 tokens]  [6 tokens] 
#>  [199] [3 tokens]  [3 tokens]  [6 tokens]  [3 tokens]  [6 tokens]  [8 tokens] 
#>  [205] [7 tokens]  [5 tokens]  [6 tokens]  [3 tokens]  [6 tokens]  [5 tokens] 
#>  [211] [5 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [3 tokens]  [3 tokens] 
#>  [217] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [7 tokens]  [7 tokens] 
#>  [223] [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens] 
#>  [229] [5 tokens]  [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [235] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [241] [19 tokens] [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens] 
#>  [247] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [253] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [259] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [3 tokens] 
#>  [265] [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [271] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [277] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [283] [3 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#>  [289] [5 tokens]  [3 tokens]  [6 tokens]  [7 tokens]  [7 tokens]  [6 tokens] 
#>  [295] [6 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#>  [301] [4 tokens]  [4 tokens]  [4 tokens]  [6 tokens]  [6 tokens]  [10 tokens]
#>  [307] [3 tokens]  [3 tokens]  [4 tokens]  [4 tokens]  [7 tokens]  [3 tokens] 
#>  [313] [3 tokens]  [6 tokens]  [1 tokens]  [14 tokens] [14 tokens] [11 tokens]
#>  [319] [9 tokens]  [6 tokens]  [9 tokens]  [4 tokens]  [11 tokens] [4 tokens] 
#>  [325] [14 tokens] [3 tokens]  [6 tokens]  [6 tokens]  [4 tokens]  [4 tokens] 
#>  [331] [4 tokens]  [4 tokens]  [6 tokens]  [12 tokens] [4 tokens]  [4 tokens] 
#>  [337] [8 tokens]  [9 tokens]  [11 tokens] [5 tokens]  [3 tokens]  [3 tokens] 
#>  [343] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [349] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [5 tokens] 
#>  [355] [3 tokens]  [4 tokens]  [13 tokens] [11 tokens] [3 tokens]  [3 tokens] 
#>  [361] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [367] [1 tokens]  [1 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [8 tokens] 
#>  [373] [6 tokens]  [5 tokens]  [9 tokens]  [6 tokens]  [5 tokens]  [5 tokens] 
#>  [379] [5 tokens]  [5 tokens]  [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [385] [3 tokens]  [3 tokens]  [2 tokens]  [7 tokens]  [7 tokens]  [7 tokens] 
#>  [391] [4 tokens]  [4 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#>  [397] [5 tokens]  [5 tokens]  [5 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#>  [403] [6 tokens]  [6 tokens]  [5 tokens]  [6 tokens]  [7 tokens]  [10 tokens]
#>  [409] [6 tokens]  [2 tokens]  [8 tokens]  [3 tokens]  [5 tokens]  [4 tokens] 
#>  [415] [2 tokens]  [17 tokens] [18 tokens] [4 tokens]  [3 tokens]  [3 tokens] 
#>  [421] [3 tokens]  [10 tokens] [3 tokens]  [4 tokens]  [3 tokens]  [3 tokens] 
#>  [427] [11 tokens] [7 tokens]  [5 tokens]  [7 tokens]  [3 tokens]  [5 tokens] 
#>  [433] [7 tokens]  [5 tokens]  [5 tokens]  [7 tokens]  [5 tokens]  [5 tokens] 
#>  [439] [5 tokens]  [7 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [7 tokens] 
#>  [445] [5 tokens]  [8 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#>  [451] [6 tokens]  [8 tokens]  [8 tokens]  [9 tokens]  [6 tokens]  [6 tokens] 
#>  [457] [6 tokens]  [6 tokens]  [4 tokens]  [3 tokens]  [6 tokens]  [6 tokens] 
#>  [463] [4 tokens]  [4 tokens]  [1 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [469] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [475] [3 tokens]  [3 tokens]  [1 tokens]  [9 tokens]  [3 tokens]  [3 tokens] 
#>  [481] [4 tokens]  [5 tokens]  [6 tokens]  [6 tokens]  [8 tokens]  [10 tokens]
#>  [487] [3 tokens]  [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [493] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [499] [3 tokens]  [3 tokens]  [4 tokens]  [3 tokens]  [8 tokens]  [15 tokens]
#>  [505] [13 tokens] [18 tokens] [12 tokens] [12 tokens] [3 tokens]  [6 tokens] 
#>  [511] [6 tokens]  [6 tokens]  [11 tokens] [10 tokens] [3 tokens]  [13 tokens]
#>  [517] [5 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [10 tokens] [9 tokens] 
#>  [523] [14 tokens] [4 tokens]  [8 tokens]  [6 tokens]  [4 tokens]  [3 tokens] 
#>  [529] [4 tokens]  [4 tokens]  [4 tokens]  [3 tokens]  [14 tokens] [8 tokens] 
#>  [535] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [8 tokens] 
#>  [541] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [547] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [4 tokens] 
#>  [553] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [559] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [12 tokens] [6 tokens] 
#>  [565] [3 tokens]  [6 tokens]  [6 tokens]  [1 tokens]  [4 tokens]  [7 tokens] 
#>  [571] [7 tokens]  [3 tokens]  [5 tokens]  [3 tokens]  [7 tokens]  [4 tokens] 
#>  [577] [9 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [583] [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [4 tokens]  [3 tokens] 
#>  [589] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [595] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [5 tokens]  [8 tokens] 
#>  [601] [3 tokens]  [2 tokens]  [10 tokens] [3 tokens]  [5 tokens]  [4 tokens] 
#>  [607] [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [17 tokens] [4 tokens] 
#>  [613] [1 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [619] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [6 tokens] 
#>  [625] [8 tokens]  [9 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [5 tokens] 
#>  [631] [3 tokens]  [4 tokens]  [3 tokens]  [5 tokens]  [7 tokens]  [6 tokens] 
#>  [637] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [5 tokens]  [2 tokens] 
#>  [643] [7 tokens]  [6 tokens]  [7 tokens]  [7 tokens]  [3 tokens]  [3 tokens] 
#>  [649] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [655] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [661] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [667] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [673] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens] 
#>  [679] [7 tokens]  [8 tokens]  [7 tokens]  [6 tokens]  [6 tokens]  [4 tokens] 
#>  [685] [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [691] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [697] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [703] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [7 tokens] 
#>  [709] [5 tokens]  [5 tokens]  [5 tokens]  [7 tokens]  [10 tokens] [11 tokens]
#>  [715] [9 tokens]  [4 tokens]  [4 tokens]  [7 tokens]  [3 tokens]  [3 tokens] 
#>  [721] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [8 tokens]  [6 tokens] 
#>  [727] [17 tokens] [10 tokens] [15 tokens] [3 tokens]  [3 tokens]  [3 tokens] 
#>  [733] [3 tokens]  [7 tokens]  [3 tokens]  [12 tokens] [2 tokens]  [4 tokens] 
#>  [739] [6 tokens]  [8 tokens]  [11 tokens] [4 tokens]  [7 tokens]  [11 tokens]
#>  [745] [13 tokens] [4 tokens]  [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens] 
#>  [751] [3 tokens]  [3 tokens]  [4 tokens]  [6 tokens]  [8 tokens]  [6 tokens] 
#>  [757] [1 tokens]  [5 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [8 tokens] 
#>  [763] [11 tokens] [11 tokens] [11 tokens] [5 tokens]  [5 tokens]  [5 tokens] 
#>  [769] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens] 
#>  [775] [3 tokens]  [3 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [781] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [787] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [793] [3 tokens]  [4 tokens]  [10 tokens] [8 tokens]  [8 tokens]  [8 tokens] 
#>  [799] [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens] 
#>  [805] [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [6 tokens]  [14 tokens]
#>  [811] [5 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [817] [3 tokens]  [3 tokens]  [1 tokens]  [9 tokens]  [4 tokens]  [3 tokens] 
#>  [823] [6 tokens]  [11 tokens] [10 tokens] [11 tokens] [4 tokens]  [3 tokens] 
#>  [829] [6 tokens]  [3 tokens]  [5 tokens]  [8 tokens]  [8 tokens]  [7 tokens] 
#>  [835] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#>  [841] [6 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [9 tokens]  [6 tokens] 
#>  [847] [9 tokens]  [8 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens] 
#>  [853] [3 tokens]  [3 tokens]  [3 tokens]  [2 tokens]  [9 tokens]  [4 tokens] 
#>  [859] [4 tokens]  [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [865] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [871] [8 tokens]  [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [877] [3 tokens]  [5 tokens]  [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [883] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [3 tokens]  [3 tokens] 
#>  [889] [3 tokens]  [11 tokens] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [895] [3 tokens]  [4 tokens]  [6 tokens]  [3 tokens]  [7 tokens]  [1 tokens] 
#>  [901] [1 tokens]  [7 tokens]  [4 tokens]  [5 tokens]  [4 tokens]  [4 tokens] 
#>  [907] [4 tokens]  [5 tokens]  [3 tokens]  [3 tokens]  [13 tokens] [3 tokens] 
#>  [913] [4 tokens]  [10 tokens] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [919] [3 tokens]  [5 tokens]  [8 tokens]  [3 tokens]  [6 tokens]  [11 tokens]
#>  [925] [7 tokens]  [1 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [5 tokens] 
#>  [931] [4 tokens]  [10 tokens] [4 tokens]  [3 tokens]  [8 tokens]  [3 tokens] 
#>  [937] [8 tokens]  [8 tokens]  [4 tokens]  [4 tokens]  [13 tokens] [3 tokens] 
#>  [943] [7 tokens]  [5 tokens]  [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [949] [4 tokens]  [6 tokens]  [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [955] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [961] [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [5 tokens]  [1 tokens] 
#>  [967] [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [6 tokens] 
#>  [973] [4 tokens]  [6 tokens]  [5 tokens]  [5 tokens]  [7 tokens]  [5 tokens] 
#>  [979] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [985] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#>  [991] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [7 tokens]  [2 tokens] 
#>  [997] [3 tokens]  [6 tokens]  [4 tokens]  [7 tokens]  [6 tokens]  [5 tokens] 
#> [1003] [7 tokens]  [6 tokens]  [2 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [1009] [5 tokens]  [5 tokens]  [4 tokens]  [9 tokens]  [3 tokens]  [3 tokens] 
#> [1015] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1021] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1027] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1033] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1039] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1045] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [3 tokens] 
#> [1051] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1057] [3 tokens]  [8 tokens]  [5 tokens]  [5 tokens]  [9 tokens]  [9 tokens] 
#> [1063] [11 tokens] [11 tokens] [11 tokens] [5 tokens]  [5 tokens]  [5 tokens] 
#> [1069] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [1075] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [1081] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [1087] [5 tokens]  [9 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [11 tokens]
#> [1093] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [22 tokens] [6 tokens] 
#> [1099] [6 tokens]  [4 tokens]  [4 tokens]  [5 tokens]  [5 tokens]  [1 tokens] 
#> [1105] [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1111] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1117] [5 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [3 tokens]  [3 tokens] 
#> [1123] [5 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [3 tokens]  [5 tokens] 
#> [1129] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1135] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1141] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1147] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1153] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [3 tokens] 
#> [1159] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1165] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1171] [3 tokens]  [3 tokens]  [4 tokens]  [1 tokens]  [14 tokens] [7 tokens] 
#> [1177] [2 tokens]  [9 tokens]  [5 tokens]  [11 tokens] [3 tokens]  [3 tokens] 
#> [1183] [3 tokens]  [12 tokens] [3 tokens]  [7 tokens]  [3 tokens]  [4 tokens] 
#> [1189] [3 tokens]  [6 tokens]  [3 tokens]  [5 tokens]  [4 tokens]  [4 tokens] 
#> [1195] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1201] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1207] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1213] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1219] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1225] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1231] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1237] [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1243] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1249] [3 tokens]  [5 tokens]  [6 tokens]  [6 tokens]  [4 tokens]  [8 tokens] 
#> [1255] [8 tokens]  [8 tokens]  [8 tokens]  [6 tokens]  [3 tokens]  [4 tokens] 
#> [1261] [9 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1267] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1273] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1279] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1285] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1291] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1297] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1303] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1309] [3 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1315] [4 tokens]  [4 tokens]  [4 tokens]  [11 tokens] [4 tokens]  [3 tokens] 
#> [1321] [3 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [7 tokens]  [4 tokens] 
#> [1327] [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1333] [3 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [9 tokens]  [6 tokens] 
#> [1339] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [1345] [5 tokens]  [5 tokens]  [7 tokens]  [6 tokens]  [4 tokens]  [4 tokens] 
#> [1351] [7 tokens]  [4 tokens]  [4 tokens]  [7 tokens]  [5 tokens]  [9 tokens] 
#> [1357] [9 tokens]  [9 tokens]  [6 tokens]  [6 tokens]  [5 tokens]  [5 tokens] 
#> [1363] [6 tokens]  [5 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [5 tokens] 
#> [1369] [6 tokens]  [4 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [1375] [4 tokens]  [5 tokens]  [5 tokens]  [8 tokens]  [7 tokens]  [4 tokens] 
#> [1381] [5 tokens]  [15 tokens] [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1387] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens] 
#> [1393] [3 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [4 tokens] 
#> [1399] [4 tokens]  [7 tokens]  [16 tokens] [4 tokens]  [7 tokens]  [5 tokens] 
#> [1405] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [1411] [5 tokens]  [5 tokens]  [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1417] [3 tokens]  [3 tokens]  [3 tokens]  [8 tokens]  [4 tokens]  [5 tokens] 
#> [1423] [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [4 tokens] 
#> [1429] [5 tokens]  [4 tokens]  [8 tokens]  [1 tokens]  [4 tokens]  [4 tokens] 
#> [1435] [4 tokens]  [13 tokens] [9 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1441] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [5 tokens] 
#> [1447] [3 tokens]  [5 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [3 tokens] 
#> [1453] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [1459] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [10 tokens] [10 tokens]
#> [1465] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens] 
#> [1471] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1477] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [8 tokens]  [8 tokens] 
#> [1483] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [5 tokens] 
#> [1489] [13 tokens] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1495] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [2 tokens] 
#> [1501] [2 tokens]  [3 tokens]  [2 tokens]  [2 tokens]  [2 tokens]  [2 tokens] 
#> [1507] [11 tokens] [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1513] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1519] [3 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1525] [3 tokens]  [1 tokens]  [7 tokens]  [5 tokens]  [5 tokens]  [3 tokens] 
#> [1531] [5 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [3 tokens]  [5 tokens] 
#> [1537] [5 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [7 tokens]  [7 tokens] 
#> [1543] [4 tokens]  [8 tokens]  [5 tokens]  [5 tokens]  [6 tokens]  [6 tokens] 
#> [1549] [6 tokens]  [13 tokens] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1555] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [5 tokens] 
#> [1561] [5 tokens]  [8 tokens]  [9 tokens]  [8 tokens]  [8 tokens]  [9 tokens] 
#> [1567] [7 tokens]  [8 tokens]  [10 tokens] [8 tokens]  [6 tokens]  [8 tokens] 
#> [1573] [6 tokens]  [6 tokens]  [6 tokens]  [7 tokens]  [6 tokens]  [6 tokens] 
#> [1579] [6 tokens]  [6 tokens]  [6 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [1585] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [7 tokens] 
#> [1591] [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [3 tokens]  [3 tokens] 
#> [1597] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1603] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [14 tokens] [13 tokens]
#> [1609] [10 tokens] [4 tokens]  [3 tokens]  [8 tokens]  [3 tokens]  [5 tokens] 
#> [1615] [4 tokens]  [4 tokens]  [4 tokens]  [20 tokens] [7 tokens]  [3 tokens] 
#> [1621] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1627] [3 tokens]  [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens] 
#> [1633] [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [3 tokens] 
#> [1639] [3 tokens]  [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1645] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1651] [3 tokens]  [3 tokens]  [5 tokens]  [6 tokens]  [1 tokens]  [3 tokens] 
#> [1657] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1663] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1669] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [7 tokens] 
#> [1675] [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [1 tokens] 
#> [1681] [3 tokens]  [3 tokens]  [3 tokens]  [2 tokens]  [2 tokens]  [3 tokens] 
#> [1687] [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [4 tokens]  [7 tokens] 
#> [1693] [5 tokens]  [7 tokens]  [7 tokens]  [5 tokens]  [6 tokens]  [5 tokens] 
#> [1699] [6 tokens]  [5 tokens]  [7 tokens]  [6 tokens]  [5 tokens]  [3 tokens] 
#> [1705] [4 tokens]  [9 tokens]  [4 tokens]  [4 tokens]  [3 tokens]  [4 tokens] 
#> [1711] [3 tokens]  [4 tokens]  [6 tokens]  [6 tokens]  [3 tokens]  [6 tokens] 
#> [1717] [6 tokens]  [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1723] [3 tokens]  [3 tokens]  [4 tokens]  [4 tokens]  [6 tokens]  [3 tokens] 
#> [1729] [6 tokens]  [13 tokens] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1735] [3 tokens]  [3 tokens]  [6 tokens]  [4 tokens]  [6 tokens]  [6 tokens] 
#> [1741] [6 tokens]  [4 tokens]  [4 tokens]  [12 tokens] [10 tokens] [4 tokens] 
#> [1747] [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1753] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1759] [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [4 tokens]  [3 tokens] 
#> [1765] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [3 tokens] 
#> [1771] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1777] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1783] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1789] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1795] [9 tokens]  [3 tokens]  [1 tokens]  [1 tokens]  [1 tokens]  [3 tokens] 
#> [1801] [3 tokens]  [3 tokens]  [1 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1807] [3 tokens]  [3 tokens]  [5 tokens]  [3 tokens]  [3 tokens]  [4 tokens] 
#> [1813] [16 tokens] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [1819] [5 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [1825] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [1831] [6 tokens]  [10 tokens] [3 tokens]  [4 tokens]  [4 tokens]  [6 tokens] 
#> [1837] [6 tokens]  [1 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [3 tokens] 
#> [1843] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1849] [4 tokens]  [4 tokens]  [6 tokens]  [3 tokens]  [10 tokens] [8 tokens] 
#> [1855] [6 tokens]  [3 tokens]  [13 tokens] [7 tokens]  [6 tokens]  [6 tokens] 
#> [1861] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [1867] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [1873] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [1879] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [1885] [6 tokens]  [6 tokens]  [6 tokens]  [4 tokens]  [6 tokens]  [6 tokens] 
#> [1891] [4 tokens]  [4 tokens]  [3 tokens]  [6 tokens]  [6 tokens]  [3 tokens] 
#> [1897] [4 tokens]  [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1903] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1909] [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1915] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1921] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1927] [3 tokens]  [3 tokens]  [5 tokens]  [6 tokens]  [5 tokens]  [5 tokens] 
#> [1933] [3 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [1939] [3 tokens]  [5 tokens]  [5 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [1945] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [1951] [10 tokens] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1957] [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1963] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1969] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1975] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [1981] [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [1987] [4 tokens]  [1 tokens]  [4 tokens]  [6 tokens]  [6 tokens]  [3 tokens] 
#> [1993] [9 tokens]  [9 tokens]  [9 tokens]  [3 tokens]  [3 tokens]  [4 tokens] 
#> [1999] [5 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2005] [6 tokens]  [6 tokens]  [6 tokens]  [7 tokens]  [6 tokens]  [4 tokens] 
#> [2011] [2 tokens]  [2 tokens]  [5 tokens]  [3 tokens]  [3 tokens]  [8 tokens] 
#> [2017] [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens] 
#> [2023] [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [3 tokens] 
#> [2029] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2035] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [8 tokens] 
#> [2041] [9 tokens]  [9 tokens]  [9 tokens]  [3 tokens]  [7 tokens]  [2 tokens] 
#> [2047] [2 tokens]  [4 tokens]  [3 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2053] [4 tokens]  [4 tokens]  [7 tokens]  [7 tokens]  [4 tokens]  [3 tokens] 
#> [2059] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2065] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2071] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2077] [3 tokens]  [3 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2083] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2089] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [2 tokens]  [5 tokens] 
#> [2095] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2101] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2107] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2113] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2119] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2125] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2131] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2137] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2143] [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [6 tokens]  [4 tokens] 
#> [2149] [4 tokens]  [4 tokens]  [4 tokens]  [2 tokens]  [3 tokens]  [3 tokens] 
#> [2155] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2161] [4 tokens]  [4 tokens]  [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2167] [4 tokens]  [12 tokens] [10 tokens] [3 tokens]  [4 tokens]  [7 tokens] 
#> [2173] [6 tokens]  [17 tokens] [7 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2179] [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2185] [3 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens] 
#> [2191] [7 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2197] [7 tokens]  [10 tokens] [12 tokens] [3 tokens]  [5 tokens]  [6 tokens] 
#> [2203] [6 tokens]  [4 tokens]  [6 tokens]  [5 tokens]  [4 tokens]  [4 tokens] 
#> [2209] [3 tokens]  [3 tokens]  [8 tokens]  [5 tokens]  [3 tokens]  [11 tokens]
#> [2215] [10 tokens] [9 tokens]  [9 tokens]  [11 tokens] [8 tokens]  [7 tokens] 
#> [2221] [7 tokens]  [7 tokens]  [9 tokens]  [8 tokens]  [3 tokens]  [3 tokens] 
#> [2227] [18 tokens] [6 tokens]  [6 tokens]  [1 tokens]  [4 tokens]  [8 tokens] 
#> [2233] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [3 tokens]  [3 tokens] 
#> [2239] [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [3 tokens] 
#> [2245] [5 tokens]  [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2251] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2257] [3 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [2263] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [2269] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [2275] [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens] 
#> [2281] [7 tokens]  [15 tokens] [3 tokens]  [1 tokens]  [21 tokens] [10 tokens]
#> [2287] [7 tokens]  [7 tokens]  [3 tokens]  [6 tokens]  [6 tokens]  [4 tokens] 
#> [2293] [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2299] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2305] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2311] [3 tokens]  [3 tokens]  [7 tokens]  [3 tokens]  [4 tokens]  [9 tokens] 
#> [2317] [3 tokens]  [3 tokens]  [6 tokens]  [3 tokens]  [4 tokens]  [7 tokens] 
#> [2323] [8 tokens]  [2 tokens]  [10 tokens] [6 tokens]  [6 tokens]  [1 tokens] 
#> [2329] [1 tokens]  [11 tokens] [9 tokens]  [9 tokens]  [9 tokens]  [9 tokens] 
#> [2335] [9 tokens]  [4 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens] 
#> [2341] [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens] 
#> [2347] [5 tokens]  [3 tokens]  [14 tokens] [3 tokens]  [3 tokens]  [5 tokens] 
#> [2353] [10 tokens] [6 tokens]  [6 tokens]  [7 tokens]  [4 tokens]  [14 tokens]
#> [2359] [7 tokens]  [7 tokens]  [11 tokens] [11 tokens] [8 tokens]  [7 tokens] 
#> [2365] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2371] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2377] [3 tokens]  [6 tokens]  [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2383] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2389] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2395] [6 tokens]  [5 tokens]  [4 tokens]  [7 tokens]  [3 tokens]  [4 tokens] 
#> [2401] [4 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [7 tokens]  [6 tokens] 
#> [2407] [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [12 tokens] [20 tokens]
#> [2413] [3 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [2419] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [3 tokens]  [5 tokens] 
#> [2425] [5 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [5 tokens]  [3 tokens] 
#> [2431] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2437] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2443] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2449] [6 tokens]  [6 tokens]  [2 tokens]  [7 tokens]  [4 tokens]  [5 tokens] 
#> [2455] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2461] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [8 tokens] 
#> [2467] [6 tokens]  [6 tokens]  [3 tokens]  [7 tokens]  [6 tokens]  [6 tokens] 
#> [2473] [12 tokens] [3 tokens]  [6 tokens]  [6 tokens]  [17 tokens] [6 tokens] 
#> [2479] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [11 tokens]
#> [2485] [3 tokens]  [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2491] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [8 tokens]  [6 tokens] 
#> [2497] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [4 tokens] 
#> [2503] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2509] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2515] [6 tokens]  [6 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [7 tokens] 
#> [2521] [5 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2527] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2533] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2539] [3 tokens]  [3 tokens]  [3 tokens]  [9 tokens]  [9 tokens]  [3 tokens] 
#> [2545] [9 tokens]  [8 tokens]  [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2551] [3 tokens]  [1 tokens]  [4 tokens]  [3 tokens]  [4 tokens]  [8 tokens] 
#> [2557] [8 tokens]  [6 tokens]  [7 tokens]  [5 tokens]  [6 tokens]  [3 tokens] 
#> [2563] [6 tokens]  [7 tokens]  [5 tokens]  [6 tokens]  [6 tokens]  [4 tokens] 
#> [2569] [11 tokens] [3 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2575] [6 tokens]  [12 tokens] [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens] 
#> [2581] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens] 
#> [2587] [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2593] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2599] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2605] [3 tokens]  [3 tokens]  [3 tokens]  [10 tokens] [8 tokens]  [6 tokens] 
#> [2611] [4 tokens]  [4 tokens]  [2 tokens]  [3 tokens]  [6 tokens]  [6 tokens] 
#> [2617] [17 tokens] [7 tokens]  [5 tokens]  [5 tokens]  [3 tokens]  [3 tokens] 
#> [2623] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2629] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens] 
#> [2635] [4 tokens]  [6 tokens]  [6 tokens]  [5 tokens]  [6 tokens]  [6 tokens] 
#> [2641] [6 tokens]  [5 tokens]  [3 tokens]  [3 tokens]  [9 tokens]  [10 tokens]
#> [2647] [5 tokens]  [10 tokens] [6 tokens]  [3 tokens]  [6 tokens]  [4 tokens] 
#> [2653] [4 tokens]  [9 tokens]  [6 tokens]  [17 tokens] [7 tokens]  [5 tokens] 
#> [2659] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [2665] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [2671] [4 tokens]  [3 tokens]  [3 tokens]  [11 tokens] [3 tokens]  [3 tokens] 
#> [2677] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2683] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2689] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2695] [3 tokens]  [3 tokens]  [3 tokens]  [2 tokens]  [3 tokens]  [3 tokens] 
#> [2701] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2707] [3 tokens]  [3 tokens]  [5 tokens]  [4 tokens]  [4 tokens]  [17 tokens]
#> [2713] [7 tokens]  [4 tokens]  [6 tokens]  [4 tokens]  [9 tokens]  [11 tokens]
#> [2719] [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [8 tokens]  [8 tokens] 
#> [2725] [3 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens] 
#> [2731] [6 tokens]  [4 tokens]  [1 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2737] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2743] [6 tokens]  [6 tokens]  [3 tokens]  [9 tokens]  [3 tokens]  [3 tokens] 
#> [2749] [3 tokens]  [4 tokens]  [3 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2755] [6 tokens]  [10 tokens] [10 tokens] [9 tokens]  [3 tokens]  [6 tokens] 
#> [2761] [6 tokens]  [4 tokens]  [3 tokens]  [7 tokens]  [4 tokens]  [7 tokens] 
#> [2767] [5 tokens]  [5 tokens]  [5 tokens]  [7 tokens]  [5 tokens]  [9 tokens] 
#> [2773] [6 tokens]  [13 tokens] [4 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2779] [3 tokens]  [3 tokens]  [6 tokens]  [4 tokens]  [6 tokens]  [5 tokens] 
#> [2785] [7 tokens]  [18 tokens] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2791] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2797] [4 tokens]  [4 tokens]  [19 tokens] [5 tokens]  [3 tokens]  [3 tokens] 
#> [2803] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [3 tokens] 
#> [2809] [3 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens] 
#> [2815] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [1 tokens]  [4 tokens] 
#> [2821] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2827] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2833] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2839] [4 tokens]  [7 tokens]  [17 tokens] [5 tokens]  [19 tokens] [4 tokens] 
#> [2845] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2851] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2857] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2863] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2869] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2875] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2881] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [2887] [10 tokens] [5 tokens]  [13 tokens] [13 tokens] [13 tokens] [5 tokens] 
#> [2893] [13 tokens] [6 tokens]  [5 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [2899] [5 tokens]  [6 tokens]  [5 tokens]  [7 tokens]  [4 tokens]  [12 tokens]
#> [2905] [10 tokens] [10 tokens] [4 tokens]  [12 tokens] [2 tokens]  [6 tokens] 
#> [2911] [6 tokens]  [6 tokens]  [6 tokens]  [4 tokens]  [4 tokens]  [11 tokens]
#> [2917] [5 tokens]  [6 tokens]  [1 tokens]  [5 tokens]  [6 tokens]  [6 tokens] 
#> [2923] [2 tokens]  [10 tokens] [4 tokens]  [4 tokens]  [7 tokens]  [4 tokens] 
#> [2929] [7 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [3 tokens]  [3 tokens] 
#> [2935] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2941] [3 tokens]  [4 tokens]  [7 tokens]  [3 tokens]  [3 tokens]  [4 tokens] 
#> [2947] [4 tokens]  [4 tokens]  [4 tokens]  [1 tokens]  [5 tokens]  [3 tokens] 
#> [2953] [3 tokens]  [6 tokens]  [10 tokens] [6 tokens]  [6 tokens]  [6 tokens] 
#> [2959] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2965] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2971] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2977] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2983] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2989] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [2995] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3001] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3007] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3013] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3019] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3025] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3031] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3037] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3043] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3049] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3055] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3061] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3067] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3073] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3079] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3085] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3091] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3097] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3103] [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [14 tokens] [8 tokens] 
#> [3109] [1 tokens]  [13 tokens] [7 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3115] [6 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [5 tokens]  [7 tokens] 
#> [3121] [7 tokens]  [5 tokens]  [15 tokens] [9 tokens]  [5 tokens]  [4 tokens] 
#> [3127] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens] 
#> [3133] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [1 tokens] 
#> [3139] [16 tokens] [17 tokens] [6 tokens]  [3 tokens]  [3 tokens]  [5 tokens] 
#> [3145] [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [4 tokens] 
#> [3151] [3 tokens]  [3 tokens]  [12 tokens] [5 tokens]  [6 tokens]  [7 tokens] 
#> [3157] [7 tokens]  [7 tokens]  [5 tokens]  [6 tokens]  [5 tokens]  [9 tokens] 
#> [3163] [6 tokens]  [12 tokens] [4 tokens]  [6 tokens]  [5 tokens]  [5 tokens] 
#> [3169] [5 tokens]  [11 tokens] [1 tokens]  [1 tokens]  [5 tokens]  [6 tokens] 
#> [3175] [6 tokens]  [3 tokens]  [3 tokens]  [9 tokens]  [10 tokens] [3 tokens] 
#> [3181] [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3187] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3193] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3199] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3205] [7 tokens]  [9 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3211] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3217] [3 tokens]  [3 tokens]  [6 tokens]  [8 tokens]  [3 tokens]  [3 tokens] 
#> [3223] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3229] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [3 tokens] 
#> [3235] [3 tokens]  [3 tokens]  [3 tokens]  [7 tokens]  [6 tokens]  [3 tokens] 
#> [3241] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3247] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3253] [3 tokens]  [5 tokens]  [11 tokens] [7 tokens]  [3 tokens]  [3 tokens] 
#> [3259] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3265] [8 tokens]  [3 tokens]  [4 tokens]  [4 tokens]  [6 tokens]  [6 tokens] 
#> [3271] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [8 tokens] 
#> [3277] [3 tokens]  [4 tokens]  [9 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [3283] [4 tokens]  [4 tokens]  [6 tokens]  [1 tokens]  [5 tokens]  [6 tokens] 
#> [3289] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [3 tokens] 
#> [3295] [7 tokens]  [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3301] [3 tokens]  [3 tokens]  [7 tokens]  [6 tokens]  [8 tokens]  [5 tokens] 
#> [3307] [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens] 
#> [3313] [9 tokens]  [9 tokens]  [4 tokens]  [3 tokens]  [9 tokens]  [6 tokens] 
#> [3319] [6 tokens]  [4 tokens]  [3 tokens]  [1 tokens]  [5 tokens]  [5 tokens] 
#> [3325] [5 tokens]  [5 tokens]  [5 tokens]  [5 tokens]  [7 tokens]  [1 tokens] 
#> [3331] [4 tokens]  [10 tokens] [3 tokens]  [9 tokens]  [15 tokens] [15 tokens]
#> [3337] [12 tokens] [8 tokens]  [5 tokens]  [19 tokens] [4 tokens]  [4 tokens] 
#> [3343] [7 tokens]  [7 tokens]  [7 tokens]  [20 tokens] [2 tokens]  [4 tokens] 
#> [3349] [5 tokens]  [9 tokens]  [4 tokens]  [5 tokens]  [6 tokens]  [6 tokens] 
#> [3355] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3361] [6 tokens]  [6 tokens]  [7 tokens]  [7 tokens]  [4 tokens]  [4 tokens] 
#> [3367] [4 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3373] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [3 tokens]  [6 tokens] 
#> [3379] [3 tokens]  [5 tokens]  [3 tokens]  [3 tokens]  [16 tokens] [5 tokens] 
#> [3385] [3 tokens]  [8 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3391] [3 tokens]  [3 tokens]  [3 tokens]  [10 tokens] [6 tokens]  [7 tokens] 
#> [3397] [19 tokens] [4 tokens]  [6 tokens]  [7 tokens]  [7 tokens]  [11 tokens]
#> [3403] [9 tokens]  [9 tokens]  [9 tokens]  [8 tokens]  [9 tokens]  [8 tokens] 
#> [3409] [4 tokens]  [11 tokens] [4 tokens]  [3 tokens]  [3 tokens]  [5 tokens] 
#> [3415] [4 tokens]  [6 tokens]  [5 tokens]  [6 tokens]  [7 tokens]  [15 tokens]
#> [3421] [7 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens] 
#> [3427] [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [4 tokens]  [4 tokens] 
#> [3433] [1 tokens]  [3 tokens]  [15 tokens] [10 tokens] [4 tokens]  [3 tokens] 
#> [3439] [5 tokens]  [4 tokens]  [4 tokens]  [7 tokens]  [4 tokens]  [4 tokens] 
#> [3445] [11 tokens] [6 tokens]  [10 tokens] [7 tokens]  [4 tokens]  [6 tokens] 
#> [3451] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [3457] [5 tokens]  [3 tokens]  [4 tokens]  [6 tokens]  [15 tokens] [7 tokens] 
#> [3463] [3 tokens]  [3 tokens]  [4 tokens]  [10 tokens] [4 tokens]  [3 tokens] 
#> [3469] [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [5 tokens]  [5 tokens] 
#> [3475] [5 tokens]  [1 tokens]  [6 tokens]  [3 tokens]  [4 tokens]  [6 tokens] 
#> [3481] [7 tokens]  [4 tokens]  [4 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3487] [6 tokens]  [5 tokens]  [3 tokens]  [19 tokens] [13 tokens] [2 tokens] 
#> [3493] [11 tokens] [6 tokens]  [6 tokens]  [3 tokens]  [10 tokens] [10 tokens]
#> [3499] [2 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [5 tokens] 
#> [3505] [5 tokens]  [5 tokens]  [20 tokens] [5 tokens]  [8 tokens]  [9 tokens] 
#> [3511] [16 tokens] [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [3 tokens] 
#> [3517] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [8 tokens] 
#> [3523] [10 tokens] [7 tokens]  [20 tokens] [14 tokens] [4 tokens]  [4 tokens] 
#> [3529] [4 tokens]  [16 tokens] [8 tokens]  [10 tokens] [16 tokens] [9 tokens] 
#> [3535] [4 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [1 tokens] 
#> [3541] [11 tokens] [11 tokens] [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3547] [4 tokens]  [5 tokens]  [4 tokens]  [4 tokens]  [15 tokens] [15 tokens]
#> [3553] [15 tokens] [15 tokens] [8 tokens]  [8 tokens]  [13 tokens] [1 tokens] 
#> [3559] [5 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [6 tokens] 
#> [3565] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3571] [4 tokens]  [4 tokens]  [4 tokens]  [2 tokens]  [3 tokens]  [6 tokens] 
#> [3577] [8 tokens]  [5 tokens]  [6 tokens]  [6 tokens]  [8 tokens]  [9 tokens] 
#> [3583] [7 tokens]  [2 tokens]  [2 tokens]  [2 tokens]  [4 tokens]  [3 tokens] 
#> [3589] [4 tokens]  [6 tokens]  [10 tokens] [6 tokens]  [6 tokens]  [6 tokens] 
#> [3595] [6 tokens]  [5 tokens]  [10 tokens] [7 tokens]  [14 tokens] [11 tokens]
#> [3601] [11 tokens] [3 tokens]  [5 tokens]  [8 tokens]  [4 tokens]  [16 tokens]
#> [3607] [8 tokens]  [19 tokens] [8 tokens]  [8 tokens]  [5 tokens]  [4 tokens] 
#> [3613] [5 tokens]  [5 tokens]  [6 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [3619] [7 tokens]  [6 tokens]  [4 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3625] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [11 tokens] [6 tokens] 
#> [3631] [5 tokens]  [5 tokens]  [6 tokens]  [4 tokens]  [4 tokens]  [6 tokens] 
#> [3637] [11 tokens] [4 tokens]  [6 tokens]  [5 tokens]  [5 tokens]  [5 tokens] 
#> [3643] [5 tokens]  [5 tokens]  [5 tokens]  [8 tokens]  [4 tokens]  [4 tokens] 
#> [3649] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [3655] [6 tokens]  [3 tokens]  [2 tokens]  [5 tokens]  [18 tokens] [17 tokens]
#> [3661] [18 tokens] [6 tokens]  [8 tokens]  [8 tokens]  [3 tokens]  [3 tokens] 
#> [3667] [4 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [6 tokens]  [6 tokens] 
#> [3673] [17 tokens] [8 tokens]  [5 tokens]  [7 tokens]  [6 tokens]  [8 tokens] 
#> [3679] [13 tokens] [10 tokens] [4 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3685] [8 tokens]  [10 tokens] [8 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [3691] [4 tokens]  [4 tokens]  [4 tokens]  [8 tokens]  [8 tokens]  [5 tokens] 
#> [3697] [6 tokens]  [9 tokens]  [9 tokens]  [9 tokens]  [5 tokens]  [8 tokens] 
#> [3703] [8 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3709] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3715] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [19 tokens] [5 tokens] 
#> [3721] [8 tokens]  [11 tokens] [4 tokens]  [3 tokens]  [3 tokens]  [11 tokens]
#> [3727] [4 tokens]  [9 tokens]  [9 tokens]  [4 tokens]  [7 tokens]  [11 tokens]
#> [3733] [4 tokens]  [8 tokens]  [4 tokens]  [8 tokens]  [5 tokens]  [6 tokens] 
#> [3739] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3745] [6 tokens]  [3 tokens]  [3 tokens]  [7 tokens]  [5 tokens]  [5 tokens] 
#> [3751] [5 tokens]  [5 tokens]  [4 tokens]  [1 tokens]  [8 tokens]  [4 tokens] 
#> [3757] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [11 tokens]
#> [3763] [5 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [3769] [4 tokens]  [4 tokens]  [3 tokens]  [12 tokens] [8 tokens]  [8 tokens] 
#> [3775] [6 tokens]  [11 tokens] [4 tokens]  [14 tokens] [3 tokens]  [3 tokens] 
#> [3781] [12 tokens] [4 tokens]  [6 tokens]  [3 tokens]  [13 tokens] [3 tokens] 
#> [3787] [3 tokens]  [5 tokens]  [8 tokens]  [12 tokens] [4 tokens]  [4 tokens] 
#> [3793] [19 tokens] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [3799] [4 tokens]  [8 tokens]  [8 tokens]  [4 tokens]  [5 tokens]  [5 tokens] 
#> [3805] [6 tokens]  [3 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3811] [5 tokens]  [5 tokens]  [4 tokens]  [6 tokens]  [4 tokens]  [4 tokens] 
#> [3817] [7 tokens]  [6 tokens]  [7 tokens]  [4 tokens]  [9 tokens]  [7 tokens] 
#> [3823] [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens]  [7 tokens] 
#> [3829] [7 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3835] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3841] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [5 tokens]  [6 tokens] 
#> [3847] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3853] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3859] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3865] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3871] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3877] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3883] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3889] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3895] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3901] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3907] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3913] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3919] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3925] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3931] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3937] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [3943] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [4 tokens] 
#> [3949] [4 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens] 
#> [3955] [1 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [6 tokens]  [6 tokens] 
#> [3961] [6 tokens]  [10 tokens] [9 tokens]  [3 tokens]  [4 tokens]  [3 tokens] 
#> [3967] [11 tokens] [4 tokens]  [8 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [3973] [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [2 tokens] 
#> [3979] [7 tokens]  [8 tokens]  [3 tokens]  [6 tokens]  [6 tokens]  [3 tokens] 
#> [3985] [3 tokens]  [8 tokens]  [6 tokens]  [7 tokens]  [4 tokens]  [7 tokens] 
#> [3991] [7 tokens]  [7 tokens]  [15 tokens] [7 tokens]  [5 tokens]  [7 tokens] 
#> [3997] [3 tokens]  [5 tokens]  [3 tokens]  [6 tokens]  [7 tokens]  [4 tokens] 
#> [4003] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [4009] [4 tokens]  [8 tokens]  [7 tokens]  [7 tokens]  [9 tokens]  [8 tokens] 
#> [4015] [5 tokens]  [4 tokens]  [3 tokens]  [6 tokens]  [7 tokens]  [4 tokens] 
#> [4021] [9 tokens]  [4 tokens]  [4 tokens]  [3 tokens]  [10 tokens] [3 tokens] 
#> [4027] [4 tokens]  [7 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [7 tokens] 
#> [4033] [4 tokens]  [5 tokens]  [3 tokens]  [10 tokens] [5 tokens]  [7 tokens] 
#> [4039] [6 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [4045] [4 tokens]  [4 tokens]  [5 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [4051] [6 tokens]  [10 tokens] [3 tokens]  [6 tokens]  [16 tokens] [3 tokens] 
#> [4057] [15 tokens] [6 tokens]  [7 tokens]  [4 tokens]  [6 tokens]  [19 tokens]
#> [4063] [4 tokens]  [10 tokens] [6 tokens]  [3 tokens]  [8 tokens]  [11 tokens]
#> [4069] [3 tokens]  [9 tokens]  [12 tokens] [8 tokens]  [17 tokens] [4 tokens] 
#> [4075] [5 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [5 tokens]  [5 tokens] 
#> [4081] [5 tokens]  [7 tokens]  [5 tokens]  [1 tokens]  [6 tokens]  [11 tokens]
#> [4087] [14 tokens] [9 tokens]  [6 tokens]  [6 tokens]  [3 tokens]  [7 tokens] 
#> [4093] [6 tokens]  [5 tokens]  [7 tokens]  [7 tokens]  [3 tokens]  [3 tokens] 
#> [4099] [4 tokens]  [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [4105] [3 tokens]  [3 tokens]  [3 tokens]  [5 tokens]  [1 tokens]  [6 tokens] 
#> [4111] [12 tokens] [11 tokens] [3 tokens]  [4 tokens]  [8 tokens]  [11 tokens]
#> [4117] [9 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [4123] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [4129] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [4135] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens] 
#> [4141] [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [10 tokens] [7 tokens] 
#> [4147] [1 tokens]  [7 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [3 tokens] 
#> [4153] [3 tokens]  [5 tokens]  [7 tokens]  [4 tokens]  [6 tokens]  [6 tokens] 
#> [4159] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [4165] [6 tokens]  [6 tokens]  [4 tokens]  [3 tokens]  [3 tokens]  [5 tokens] 
#> [4171] [3 tokens]  [3 tokens]  [3 tokens]  [16 tokens] [11 tokens] [3 tokens] 
#> [4177] [17 tokens] [5 tokens]  [5 tokens]  [5 tokens]  [4 tokens]  [4 tokens] 
#> [4183] [4 tokens]  [4 tokens]  [4 tokens]  [10 tokens] [4 tokens]  [9 tokens] 
#> [4189] [9 tokens]  [9 tokens]  [4 tokens]  [5 tokens]  [4 tokens]  [4 tokens] 
#> [4195] [8 tokens]  [9 tokens]  [4 tokens]  [1 tokens]  [4 tokens]  [5 tokens] 
#> [4201] [5 tokens]  [5 tokens]  [6 tokens]  [5 tokens]  [5 tokens]  [4 tokens] 
#> [4207] [8 tokens]  [10 tokens] [12 tokens] [4 tokens]  [9 tokens]  [15 tokens]
#> [4213] [3 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [4 tokens]  [3 tokens] 
#> [4219] [1 tokens]  [3 tokens]  [8 tokens]  [8 tokens]  [8 tokens]  [8 tokens] 
#> [4225] [8 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [4 tokens] 
#> [4231] [6 tokens]  [3 tokens]  [3 tokens]  [3 tokens]  [4 tokens]  [3 tokens] 
#> [4237] [3 tokens]  [3 tokens]  [9 tokens]  [14 tokens] [3 tokens]  [1 tokens] 
#> [4243] [7 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [4249] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [4255] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [4261] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [4267] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [4273] [7 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens] 
#> [4279] [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [6 tokens]  [4 tokens] 
#> # Unique Tokens: 952